perf: optimize ada address generation
