+++ package.json
@@ -141,11 +141,11 @@
     "@mysten/sui": "1.28.0",
     "@ngraveio/bc-ur": "^1.1.13",
     "@nktkas/hyperliquid": "0.32.2",
-    "@onekeyfe/cross-inpage-provider-core": "2.2.70",
-    "@onekeyfe/cross-inpage-provider-errors": "2.2.70",
-    "@onekeyfe/cross-inpage-provider-injected": "2.2.70",
-    "@onekeyfe/cross-inpage-provider-types": "2.2.70",
-    "@onekeyfe/extension-bridge-hosted": "2.2.70",
+    "@onekeyfe/cross-inpage-provider-core": "2.2.71",
+    "@onekeyfe/cross-inpage-provider-errors": "2.2.71",
+    "@onekeyfe/cross-inpage-provider-injected": "2.2.71",
+    "@onekeyfe/cross-inpage-provider-types": "2.2.71",
+    "@onekeyfe/extension-bridge-hosted": "2.2.71",
     "@onekeyfe/hd-ble-sdk": "1.1.27-alpha.4",
     "@onekeyfe/hd-common-connect-sdk": "1.1.27-alpha.4",
     "@onekeyfe/hd-core": "1.1.27-alpha.4",
@@ -157,7 +157,7 @@
     "@onekeyfe/hwk-ledger-adapter": "1.1.27-alpha.4",
     "@onekeyfe/hwk-ledger-connector-ble": "1.1.27-alpha.4",
     "@onekeyfe/hwk-ledger-connector-webhid": "1.1.27-alpha.4",
-    "@onekeyfe/onekey-cross-webview": "2.2.70",
+    "@onekeyfe/onekey-cross-webview": "2.2.71",
     "@polkadot/extension-inject": "0.54.1",
     "@polkadot/types": "14.3.1",
     "@polkadot/util-crypto": "13.2.3",
