+++ packages/core/src/chains/ada/sdkAda/bip32.ts
@@ -8,12 +8,154 @@ import { bech32, mnemonicToRootKeypair, toPublic } from 'cardano-crypto.js';
 import {
   type IHdCredentialDecryptCacheParams,
   mnemonicFromEntropy,
+  mnemonicToEntropy,
+  rawEntropyFromHdCredential,
 } from '@onekeyhq/core/src/secret';
+import { OneKeyLocalError } from '@onekeyhq/shared/src/errors';
+import RN_QUICK_CRYPTO from '@onekeyhq/shared/src/modules3rdParty/react-native-quick-crypto';
+import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
 import { DERIVATION_SCHEME, HARDENED_THRESHOLD } from './constants';
 
 import type { ICoreHdCredentialEncryptHex } from '../../../types';
 import type { IAdaBIP32Path } from '../types';
+
+export type IAdaRootKeyPerfTraceEvent = {
+  durationMs: number;
+  metadata?: Record<string, boolean | number | string | undefined>;
+  name:
+    | 'mnemonicFromEntropy'
+    | 'mnemonicToRootKeypair'
+    | 'mnemonicToRootKeypair.quickCryptoPbkdf2Sha512'
+    | 'mnemonicToRootKeypair.toPublic'
+    | 'rawEntropyFromHdCredential';
+};
+
+export type IAdaRootKeyPerfTrace = {
+  onEvent: (event: IAdaRootKeyPerfTraceEvent) => void;
+};
+
+type IGetRootKeyOptions = IHdCredentialDecryptCacheParams & {
+  perfTrace?: IAdaRootKeyPerfTrace;
+};
+
+function traceAdaRootKey<T>({
+  fn,
+  metadata,
+  name,
+  perfTrace,
+}: {
+  fn: () => T;
+  metadata?: IAdaRootKeyPerfTraceEvent['metadata'];
+  name: IAdaRootKeyPerfTraceEvent['name'];
+  perfTrace?: IAdaRootKeyPerfTrace;
+}): T {
+  if (!perfTrace) {
+    return fn();
+  }
+  const start = Date.now();
+  try {
+    return fn();
+  } finally {
+    perfTrace.onEvent({
+      durationMs: Date.now() - start,
+      metadata,
+      name,
+    });
+  }
+}
+
+async function traceAdaRootKeyAsync<T>({
+  fn,
+  metadata,
+  name,
+  perfTrace,
+}: {
+  fn: () => Promise<T>;
+  metadata?: IAdaRootKeyPerfTraceEvent['metadata'];
+  name: IAdaRootKeyPerfTraceEvent['name'];
+  perfTrace?: IAdaRootKeyPerfTrace;
+}): Promise<T> {
+  if (!perfTrace) {
+    return fn();
+  }
+  const start = Date.now();
+  try {
+    return await fn();
+  } finally {
+    perfTrace.onEvent({
+      durationMs: Date.now() - start,
+      metadata,
+      name,
+    });
+  }
+}
+
+function canUseQuickCryptoRootKeypairV2() {
+  return platformEnv.isNative && !platformEnv.isJest && DERIVATION_SCHEME === 2;
+}
+
+async function entropyToRootKeypairV2ByQuickCrypto({
+  entropy,
+  perfTrace,
+}: {
+  entropy: Buffer;
+  perfTrace?: IAdaRootKeyPerfTrace;
+}) {
+  const seed = Buffer.from(entropy);
+  const xprv = await traceAdaRootKeyAsync({
+    perfTrace,
+    name: 'mnemonicToRootKeypair.quickCryptoPbkdf2Sha512',
+    metadata: {
+      digest: 'sha512',
+      iterations: 4096,
+      keyLength: 96,
+    },
+    fn: async () => {
+      try {
+        return await new Promise<Buffer>((resolve, reject) => {
+          RN_QUICK_CRYPTO.pbkdf2('', seed, 4096, 96, 'sha512', (err, key) => {
+            if (err || !key) {
+              reject(
+                err ||
+                  new OneKeyLocalError(
+                    'quick-crypto ADA root key PBKDF2 failed',
+                  ),
+              );
+              return;
+            }
+            resolve(Buffer.from(key));
+          });
+        });
+      } finally {
+        seed.fill(0);
+      }
+    },
+  });
+
+  xprv[0] &= 248;
+  xprv[31] &= 31;
+  xprv[31] |= 64;
+  const publicKey = traceAdaRootKey({
+    perfTrace,
+    name: 'mnemonicToRootKeypair.toPublic',
+    fn: () => toPublic(xprv.slice(0, 64)),
+  });
+  return Buffer.concat([xprv.slice(0, 64), publicKey, xprv.slice(64)]);
+}
+
+async function mnemonicToRootKeypairV2ByQuickCrypto({
+  mnemonic,
+  perfTrace,
+}: {
+  mnemonic: string;
+  perfTrace?: IAdaRootKeyPerfTrace;
+}) {
+  return entropyToRootKeypairV2ByQuickCrypto({
+    entropy: Buffer.from(mnemonicToEntropy(mnemonic), 'hex'),
+    perfTrace,
+  });
+}
 
 export function toBip32StringPath(derivationPath: IAdaBIP32Path) {
   return `m/${derivationPath
@@ -32,14 +174,81 @@ export function getPathIndex(path: string) {
 export async function getRootKey(
   password: string,
   hdCredential: ICoreHdCredentialEncryptHex,
-  options?: IHdCredentialDecryptCacheParams,
+  options?: IGetRootKeyOptions,
 ): Promise<Buffer> {
+  if (canUseQuickCryptoRootKeypairV2()) {
+    const rawEntropyStart = Date.now();
+    const entropy = await rawEntropyFromHdCredential(
+      hdCredential,
+      password,
+      options,
+    );
+    options?.perfTrace?.onEvent({
+      durationMs: Date.now() - rawEntropyStart,
+      name: 'rawEntropyFromHdCredential',
+    });
+    try {
+      return getRootKeyFromRawEntropy(entropy, options);
+    } finally {
+      entropy.fill(0);
+    }
+  }
+
+  const mnemonicStart = Date.now();
   const mnemonic: string = await mnemonicFromEntropy(
     hdCredential,
     password,
     options,
   );
-  const rootKey = await mnemonicToRootKeypair(mnemonic, DERIVATION_SCHEME);
+  options?.perfTrace?.onEvent({
+    durationMs: Date.now() - mnemonicStart,
+    name: 'mnemonicFromEntropy',
+  });
+  return getRootKeyFromMnemonic(mnemonic, options);
+}
+
+async function getRootKeyFromRawEntropy(
+  entropy: Buffer,
+  options?: {
+    perfTrace?: IAdaRootKeyPerfTrace;
+  },
+): Promise<Buffer> {
+  const rootKeyStart = Date.now();
+  const rootKey = await entropyToRootKeypairV2ByQuickCrypto({
+    entropy,
+    perfTrace: options?.perfTrace,
+  });
+  options?.perfTrace?.onEvent({
+    durationMs: Date.now() - rootKeyStart,
+    metadata: {
+      derivationScheme: DERIVATION_SCHEME,
+      source: 'rawEntropy',
+    },
+    name: 'mnemonicToRootKeypair',
+  });
+  return rootKey;
+}
+
+export async function getRootKeyFromMnemonic(
+  mnemonic: string,
+  options?: {
+    perfTrace?: IAdaRootKeyPerfTrace;
+  },
+): Promise<Buffer> {
+  const rootKeyStart = Date.now();
+  const rootKey = canUseQuickCryptoRootKeypairV2()
+    ? await mnemonicToRootKeypairV2ByQuickCrypto({
+        mnemonic,
+        perfTrace: options?.perfTrace,
+      })
+    : await mnemonicToRootKeypair(mnemonic, DERIVATION_SCHEME);
+  options?.perfTrace?.onEvent({
+    durationMs: Date.now() - rootKeyStart,
+    metadata: {
+      derivationScheme: DERIVATION_SCHEME,
+    },
+    name: 'mnemonicToRootKeypair',
+  });
   return rootKey;
 }
