+++ packages/core/src/chains/ada/sdkAda/sdk/adaWebSdk.jest.ts
@@ -1,5 +1,7 @@
 import * as Loader from '@onekeyfe/cardano-coin-selection-asmjs';
 
+import { batchGetShelleyAddressByRootKeyHex } from '../shelley-address';
+
 import type { IGetCardanoApi } from './types';
 
 const getCardanoApi: IGetCardanoApi = async () => ({
@@ -17,6 +19,7 @@ const getCardanoApi: IGetCardanoApi = async () => ({
   parseRawTxBodyStakeInfo: Loader.onekeyUtils.parseRawTxBodyStakeInfo,
   extractStakeKeyHashFromBaseAddress:
     Loader.onekeyUtils.extractStakeKeyHashFromBaseAddress,
+  batchGetShelleyAddressByRootKeyHex,
 });
 
 export default {
