+++ packages/core/src/chains/ada/sdkAda/sdk/adaWebSdk.ts
@@ -1,3 +1,5 @@
+import { batchGetShelleyAddressByRootKeyHex } from '../shelley-address';
+
 import type { IGetCardanoApi } from './types';
 
 const getCardanoApi: IGetCardanoApi = async () => {
@@ -17,6 +19,7 @@ const getCardanoApi: IGetCardanoApi = async () => {
     parseRawTxBodyStakeInfo: Loader.onekeyUtils.parseRawTxBodyStakeInfo,
     extractStakeKeyHashFromBaseAddress:
       Loader.onekeyUtils.extractStakeKeyHashFromBaseAddress,
+    batchGetShelleyAddressByRootKeyHex,
   };
 };
