+++ packages/core/src/chains/ada/sdkAda/sdk/types.ts
@@ -1,3 +1,8 @@
+import type {
+  IAdaBatchGetShelleyAddressByRootKeyHexParams,
+  IAdaShelleyAddressInfo,
+} from '../shelley-address';
+
 export type IGetCardanoApi = () => Promise<IAdaSdkApi>;
 
 export type IEnsureSDKReady = () => Promise<boolean>;
@@ -21,4 +26,7 @@ export interface IAdaSdkApi {
   parseRawTxInputs: typeof import('@onekeyfe/cardano-coin-selection-asmjs').onekeyUtils.parseRawTxInputs;
   parseRawTxBodyStakeInfo: typeof import('@onekeyfe/cardano-coin-selection-asmjs').onekeyUtils.parseRawTxBodyStakeInfo;
   extractStakeKeyHashFromBaseAddress: typeof import('@onekeyfe/cardano-coin-selection-asmjs').onekeyUtils.extractStakeKeyHashFromBaseAddress;
+  batchGetShelleyAddressByRootKeyHex: (
+    params: IAdaBatchGetShelleyAddressByRootKeyHexParams,
+  ) => IAdaShelleyAddressInfo[] | Promise<IAdaShelleyAddressInfo[]>;
 }
