+++ packages/kit/src/views/Developer/pages/Gallery/Components/stories/CryptoGallery.tsx
@@ -1,4387 +1,20 @@
-/* eslint-disable prefer-const */
-import { useState } from 'react';
+import { SizableText, Stack } from '@onekeyhq/components';
+import { AppCryptoTestEmoji } from '@onekeyhq/shared/src/appCrypto/utils';
 
-import crypto from 'crypto';
-
-import { pbkdf2Async as nobleP2Async } from '@noble/hashes/pbkdf2';
-import { sha512 as nobleSha512 } from '@noble/hashes/sha512';
-
+import { AdaAddressPerfTest } from './CryptoGallery/AdaAddressPerfTest';
+import { AESGcmV2Test } from './CryptoGallery/AesGcmV2Test';
 import {
-  Button,
-  DebugRenderTracker,
-  Icon,
-  NATIVE_HIT_SLOP,
-  Select,
-  SizableText,
-  Stack,
-  Toast,
-  View,
-  XStack,
-  YStack,
-} from '@onekeyhq/components';
-import { useClipboard } from '@onekeyhq/components/src/hooks/useClipboard';
-import {
-  ANIMATE_ONLY_OPACITY,
-  ANIMATE_ONLY_TRANSFORM,
-} from '@onekeyhq/components/src/utils/animationConstants';
-import type {
-  IBatchGetPublicKeysPerfTrace,
-  IBatchGetPublicKeysPerfTraceEvent,
-  IBip39RevealableSeed,
-  IMnemonicToSeedPerfTrace,
-  IMnemonicToSeedPerfTraceEvent,
-} from '@onekeyhq/core/src/secret';
-import type { ICurveName } from '@onekeyhq/core/src/types';
-import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { useDemoPriceInfoAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/demo';
-import appCrypto from '@onekeyhq/shared/src/appCrypto';
-import type { IRunAppCryptoTestTaskResult } from '@onekeyhq/shared/src/appCrypto/utils';
-import {
-  AppCryptoTestEmoji,
-  runAppCryptoTestTask,
-} from '@onekeyhq/shared/src/appCrypto/utils';
-import { OneKeyLocalError } from '@onekeyhq/shared/src/errors';
-import RN_AES from '@onekeyhq/shared/src/modules3rdParty/react-native-aes-crypto';
-import RN_FAST_PBKDF2 from '@onekeyhq/shared/src/modules3rdParty/react-native-fast-pbkdf2';
-import RN_QUICK_CRYPTO from '@onekeyhq/shared/src/modules3rdParty/react-native-quick-crypto';
-import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import bufferUtils from '@onekeyhq/shared/src/utils/bufferUtils';
-import stringUtils from '@onekeyhq/shared/src/utils/stringUtils';
-
+  AESCbcTest,
+  HashTest,
+  KeyGenTest,
+  PBKDF2Test,
+} from './CryptoGallery/BasicCryptoTests';
+import { CryptoSubtlePolyfillTest } from './CryptoGallery/CryptoSubtlePolyfillTest';
+import { JotaiDemoPriceInfo } from './CryptoGallery/JotaiDemoPriceInfo';
+import { NativeWebEmbedCryptoPerfTest } from './CryptoGallery/NativeWebEmbedCryptoPerfTest';
+import { SecretFunctionsTest } from './CryptoGallery/SecretFunctionsTest';
+import { CustomAccordion, CustomAccordionItem } from './CryptoGallery/shared';
 import { Layout } from './utils/Layout';
-
-// Core secret functions are loaded dynamically to avoid kit->core value import
-async function loadCoreSecret() {
-  return import('@onekeyhq/core/src/secret');
-}
-
-function PartContainer({
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  title,
-  children,
-}: {
-  title: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <YStack>
-      <YStack gap="$5">{children}</YStack>
-    </YStack>
-  );
-}
-
-// Custom Accordion Components
-function CustomAccordion({ children }: { children: React.ReactNode }) {
-  return (
-    <YStack gap="$2" width="100%">
-      {children}
-    </YStack>
-  );
-}
-
-function CustomAccordionItem({
-  title,
-  children,
-}: {
-  title: string;
-  children: React.ReactNode;
-}) {
-  const [isOpen, setIsOpen] = useState(false);
-
-  return (
-    <YStack
-      // borderColor="$border"
-      // borderWidth={StyleSheet.hairlineWidth}
-      // borderRadius="$2"
-      overflow="hidden"
-    >
-      <Stack
-        flexDirection="row"
-        justifyContent="space-between"
-        alignItems="center"
-        py="$4"
-        px="$2"
-        borderRadius="$2"
-        pressStyle={{ opacity: 0.7 }}
-        onPress={() => setIsOpen(!isOpen)}
-        bg="$backgroundFocus"
-      >
-        <SizableText>{title}</SizableText>
-        <View
-          animation="quick"
-          animateOnly={ANIMATE_ONLY_TRANSFORM}
-          rotate={isOpen ? '0deg' : '-90deg'}
-          transformOrigin="center"
-        >
-          <Icon name="ChevronDownSmallOutline" color="$iconSubdued" size="$6" />
-        </View>
-      </Stack>
-
-      <YStack
-        animation="quick"
-        animateOnly={ANIMATE_ONLY_OPACITY}
-        opacity={isOpen ? 1 : 0}
-        overflow="hidden"
-        style={{
-          maxHeight: isOpen ? 100_000 : 0,
-          transition: 'max-height 0.3s ease-in-out',
-        }}
-      >
-        <YStack paddingTop="$2">{children}</YStack>
-      </YStack>
-    </YStack>
-  );
-}
-
-const CRYPTO_GALLERY_TEST_COOLDOWN_MS = 500;
-const CRYPTO_GALLERY_DEFAULT_PATH_EMOJI = '⭐';
-let cryptoGalleryTestQueue = Promise.resolve();
-let lastCryptoGalleryTestFinishedAt = 0;
-
-function waitForCryptoGalleryTestCooldown() {
-  const waitMs =
-    lastCryptoGalleryTestFinishedAt > 0
-      ? Math.max(
-          CRYPTO_GALLERY_TEST_COOLDOWN_MS -
-            (Date.now() - lastCryptoGalleryTestFinishedAt),
-          0,
-        )
-      : 0;
-  return new Promise<void>((resolve) => {
-    setTimeout(resolve, waitMs);
-  });
-}
-
-function runCryptoGalleryTestExclusive<T>(fn: () => Promise<T>): Promise<T> {
-  const run = cryptoGalleryTestQueue
-    .catch(() => undefined)
-    .then(async () => {
-      await waitForCryptoGalleryTestCooldown();
-      try {
-        return await fn();
-      } finally {
-        lastCryptoGalleryTestFinishedAt = Date.now();
-      }
-    });
-  cryptoGalleryTestQueue = run.then(
-    () => undefined,
-    () => undefined,
-  );
-  return run;
-}
-
-function stringifyCryptoGalleryTablePayload(value: unknown): string {
-  return stringUtils.stableStringify(
-    value,
-    stringUtils.STRINGIFY_REPLACER.bufferToHex,
-    2,
-  );
-}
-
-// Test Components
-function PBKDF2Test() {
-  const [result, setResult] = useState('');
-
-  const testPBKDF2 = async () => {
-    try {
-      const r = await appCrypto.pbkdf2.$testSampleForPbkdf2();
-      setResult(JSON.stringify(r, null, 2));
-      Toast.success({
-        title: `PBKDF2 completed `,
-      });
-    } catch (error) {
-      Toast.error({
-        title: `PBKDF2 failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="PBKDF2 Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testPBKDF2)}
-      >
-        Test PBKDF2
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-function HashTest() {
-  const [result, setResult] = useState('');
-
-  const testHash = async () => {
-    try {
-      const r = await appCrypto.hash.$testSampleForHash();
-      setResult(JSON.stringify(r, null, 2));
-      Toast.success({
-        title: `Hash completed `,
-      });
-    } catch (error) {
-      Toast.error({
-        title: `Hash failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="Hash Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testHash)}
-      >
-        Test Hash
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-function KeyGenTest() {
-  const [result, setResult] = useState('');
-
-  const testKeyGen = async () => {
-    try {
-      const r = await appCrypto.keyGen.$testSampleForKeyGen();
-      setResult(JSON.stringify(r, null, 2));
-      Toast.success({
-        title: `KeyGen completed`,
-      });
-    } catch (error) {
-      Toast.error({
-        title: `KeyGen failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="KeyGen Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testKeyGen)}
-      >
-        Test KeyGen
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-function AESCbcTest() {
-  const [result, setResult] = useState('');
-
-  const testAESCbc = async () => {
-    try {
-      const r = await appCrypto.aesCbc.$testSampleForAesCbc();
-      setResult(JSON.stringify(r, null, 2));
-      Toast.success({
-        title: `AES-CBC completed`,
-      });
-    } catch (error) {
-      Toast.error({
-        title: `AES-CBC failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="AES-CBC Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testAESCbc)}
-      >
-        Test AES-CBC
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-// Production default (600_000) is always measured automatically, so it is
-// intentionally excluded from the selector.
-const AES_GCM_V2_ITER_OPTIONS: number[] = [
-  5000, 6000, 7000, 8000, 9000, 10_000, 20_000, 50_000, 100_000, 400_000,
-  800_000, 1_000_000, 1_500_000, 2_000_000, 3_000_000, 5_000_000,
-];
-const AES_GCM_V2_DEFAULT_ITER = 600_000;
-const AES_GCM_V2_BENCHMARK_COOLDOWN_MS = CRYPTO_GALLERY_TEST_COOLDOWN_MS;
-
-// Canonical PBKDF2-HMAC-SHA256 vectors for the fixed gallery inputs:
-//   password = sha256('onekey-gallery-password')
-//            = 5251c069b517bbf3f2a4b41fef38d2470c9967c1b265c9d3261c3f48eeadaf01
-//   salt     = 303132333435363738393a3b3c3d3e3f404142434445464748494a4b4c4d4e4f
-//   dkLen    = 32, hash = sha256
-// Cross-verified with Node's `crypto.pbkdf2Sync` and `@noble/hashes/pbkdf2`
-// (two independent implementations). Pinning these forces every backend on
-// every platform to match a third-party reference, not just each other.
-const AES_GCM_V2_PBKDF2_EXPECTED: Record<number, string> = {
-  5000: 'a0b5a344d633782e09c3b12c25e7af2f7884046d3b6532fbf8703621f8fa85fb',
-  6000: '0d68b34a893246887ca00c4f0f742fee31b12e59dfb6f467798487e745abbc35',
-  7000: 'b59e473f7f8be89d98dfbf3ac361ec57cf6504afec911bf5c6a1b8041fc96aad',
-  8000: '1122868ea2381f56dd1f28941660ea1bb4408480692aa5481177e4185d25ad2c',
-  9000: 'f5a430e87254d0085fca9a362772e6182f3559979919184d97e2bb650c45c2f5',
-  10_000: '50dd37fe740120de701248818e084bc4752aa72a4c7eece93bff4b387450144b',
-  20_000: '7029da313d766845fb9ae8004bb25d13a3d3043499e3dfee673286f4ff505290',
-  50_000: '25bc707d3b8581c4468b3de4f85eee7d860b3bd932a4530aca1b907b35326f55',
-  100_000: 'a86fa53cac84f4f57f18afef80a7eb4f178df47c1e7e1fe6310745fdf44ce570',
-  400_000: 'cff1d771a919abdaff4b5c5f03e6e2dbfdc14103440b76f35169a7536a23da8a',
-  600_000: 'b9bb3c9fd4a62196b2c4e3610bc1c92b834cdf2b1d688675e46b711f7ffd9f0a',
-  800_000: '438761b9323817c39a3a61487e164704a5d0b33ddba7bb0679d481269b712fb2',
-  1_000_000: 'bc603304bdc4e26aec16bb99cafeb780ff80440451ddc945540b87e3c66d2e3c',
-  1_500_000: '0bee8eca4c39ff15f1eb050aca8e27f3a1ee6d476921f236c47d1b8b5d845237',
-  2_000_000: 'bd36de07b6fa4a981d0e283f8ad127b0e3fcfd34a87a91634adccb89ed47959c',
-  3_000_000: '1c4dd6896f4a795132e9857c8bbc09bcc63828247a8aec45b7c259dd103a4ab3',
-  5_000_000: 'd76d404606138a9ff379e93952b7b49141cecd8146b87214bc00f82ebf80b094',
-};
-
-// PBKDF2-HMAC-SHA512, dkLen = 64 (TON / BIP39 path). Same password + salt as
-// the sha-256 vectors above. Cross-verified with Node `crypto.pbkdf2Sync` and
-// `@noble/hashes/pbkdf2` + `@noble/hashes/sha512`. This protects the
-// `@ton/crypto-primitives` -> `react-native-fast-pbkdf2.derive(... 'sha-512')`
-// chain that all TON mnemonic flows depend on at runtime.
-const AES_GCM_V2_PBKDF2_SHA512_KEY_LENGTH = 64;
-const AES_GCM_V2_PBKDF2_SHA512_EXPECTED: Record<number, string> = {
-  5000: 'cd3c233aa936e7c60ab3c0b4805c520cddd9252923d407751db75b8ef73af0ca31806f29dda7d0b88693ae2571f2bee8fb6054c284521c75a2e8aee20d262c80',
-  6000: '6c69912d9b96d5b4ab357b98d2100784b3381ddac6d3785444476af3dffa8faa8cbbb8d417092b531105afdab8e15a6d2ee210b1cff4964bbc2bb6db6690e16f',
-  7000: 'f3af99963b65c8d6af7b49af857bf014a545778255e62d9ce1b3635be05ca5d36f2497a52ff9e15b0c170e467563de773d6914749c78e0efc568876953b49938',
-  8000: 'f814e1dcf7c42d32d7e5364809c04a791cdc321957e5145925fb3fbc3facbd83728b01cffca5296e6a1d1038490e96f57a2572d1b3a2c927fff91f114504e13b',
-  9000: '9b07bbcaa4d18e728c27a2d52aacb0337b492b93ed49c9dc47f6bfcae9cd1603c2ce82b93469698f2ff599acb3ba7d56f30f705d27a0fa8e798edc40aa4584c1',
-  10_000:
-    '1f7224c0ab55f14d7a00868f40163d2ea5334487c935ca5bee77c5ab62c362ac218a7386c129d89c3a188ff117d18855fb88d7e08d56e7db7ee3d795a39d2a9b',
-  20_000:
-    '8e8fe967dfe7bf8de32f0722e33ecbde6b2d0fc5e89d890733be3e3408c755e89b26d2529669b231a4e0e2f44f46e5d6d59f77fe96bbe7c5a62a8bf50f6f16b7',
-  50_000:
-    '7b1acb1ea03ed97dd0dcadef1c2d0d0dfc6171a226e18788a429e4840a507244e3dac2af2d87e542e15751f12a6b5762c2ca80a39a981aa2e166fdfc04fe86c0',
-  100_000:
-    '13bf3c7443ebb865f13c5ccf094d273ab250c25f1d66d2828700628dce883c18def3b4578b558f8502c7e6f86b253dfacdacd86a224f1d893f4368e531f53315',
-  400_000:
-    'e137c79d100cec230f8ee30912bc775ecee64826255dc3cf17f2d54b362bace69332724986010656c8aa4efdc8d1f8d40f506f1d4b13d93b44a10898cf38ea7d',
-  600_000:
-    '61c136276cff0706297c8aa9c7eeae5ae1e9743668ec08cfe2edcd858cdf0e861609d39d853c24092da4273b34faa8e7836c7b16969d1563907f6507500a2614',
-  800_000:
-    '4fef7c43ad150b53e598a2f8b5876ad4b0080428afe7d90c3f3cf7a1fb5aca79ecfb34fa1dce4c3efb9f45c7320ab34dc2f51f356479f280c32ab0ef692bc63b',
-  1_000_000:
-    'a0b63d29e5b19788b6cfbbb56300cab23bfcb4b9ea0a0c0890bffab67764a7b286c660721ce420563c79e90e27a9231baa85de031a6ad25260115abf0d834f1f',
-  1_500_000:
-    '8a2c13a5d4a36e8d726b14ea763e8739659a4ec32919d6b20d46ec0e39d62babcc3ca7bdac619893608c5cbe8650f1c16baafdf91771082c664b793b97b40cc8',
-  2_000_000:
-    '906578898bb3e45bbe173679d50ecdd32163339bbe21604991c66dca61ce31e9439e28dcf3b7687889a9194388c1c69cf9ef21ae4d680f64fc6035cadde1f5e4',
-  3_000_000:
-    '8597c425d12f6665a3c4fcfccf3d1528eec91a9d96bba1054951c1d1945088e5b5844250834e17a85adf3cefcdac34043bb3b85a21b5b2e770918a55ac2c74cf',
-  5_000_000:
-    '181ddb90f4af03a10870271ec9bf7ee32aa304c42facb8663702de3a8c140a24cdd846b7a659e7476d71b58eb541c0e7382c42cfece96c9a2b53723cab7f2948',
-};
-
-type IPbkdf2ActualOutputs = {
-  iter: number;
-  sha256: {
-    defaultBackend?: string;
-    fastNative?: string;
-    native?: string;
-    noble?: string;
-    quickNative?: string;
-  };
-  sha512: {
-    fastNative?: string;
-    native?: string;
-    noble?: string;
-    quickNative?: string;
-  };
-};
-
-async function pbkdf2Sha512ByRNAes(
-  passwordHex: string,
-  saltHex: string,
-  iterations: number,
-): Promise<string> {
-  const key = await RN_AES.pbkdf2(
-    passwordHex,
-    saltHex,
-    iterations,
-    AES_GCM_V2_PBKDF2_SHA512_KEY_LENGTH * 8,
-    'sha512',
-  );
-  return key;
-}
-
-async function pbkdf2Sha512ByRNFast(
-  passwordBytes: Buffer,
-  saltBytes: Buffer,
-  iterations: number,
-): Promise<string> {
-  const key = await RN_FAST_PBKDF2.derive(
-    bufferUtils.bytesToBase64(passwordBytes),
-    bufferUtils.bytesToBase64(saltBytes),
-    iterations,
-    AES_GCM_V2_PBKDF2_SHA512_KEY_LENGTH,
-    'sha-512',
-  );
-  return bufferUtils.bytesToHex(bufferUtils.base64ToBytes(key));
-}
-
-async function pbkdf2Sha512ByRNQuickCrypto(
-  passwordBytes: Buffer,
-  saltBytes: Buffer,
-  iterations: number,
-): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-    RN_QUICK_CRYPTO.pbkdf2(
-      passwordBytes,
-      saltBytes,
-      iterations,
-      AES_GCM_V2_PBKDF2_SHA512_KEY_LENGTH,
-      'sha512',
-      (err, key) => {
-        if (err || !key) {
-          reject(
-            err || new OneKeyLocalError('quick-crypto PBKDF2-SHA512 failed'),
-          );
-          return;
-        }
-        resolve(bufferUtils.bytesToHex(Buffer.from(key)));
-      },
-    );
-  });
-}
-
-async function pbkdf2Sha512ByNoble(
-  passwordBytes: Buffer,
-  saltBytes: Buffer,
-  iterations: number,
-): Promise<string> {
-  const key = await nobleP2Async(nobleSha512, passwordBytes, saltBytes, {
-    c: iterations,
-    dkLen: AES_GCM_V2_PBKDF2_SHA512_KEY_LENGTH,
-  });
-  return bufferUtils.bytesToHex(Buffer.from(key));
-}
-
-function waitForAesGcmV2BenchmarkCooldown() {
-  return new Promise<void>((resolve) => {
-    setTimeout(resolve, AES_GCM_V2_BENCHMARK_COOLDOWN_MS);
-  });
-}
-
-type IAesGcmV2TableRow = {
-  opName: string;
-  actualDefaultAesGcmBackend: IAesGcmV2TableBackend | null;
-  iter: number | null;
-  actualDefaultBackend: IAesGcmV2TableBackend | null;
-  asmcrypto: number | undefined;
-  fastNative: number | undefined;
-  isActualDefaultIter: boolean;
-  noble: number | undefined;
-  native: number | undefined;
-  quickNative: number | undefined;
-  webcrypto: number | undefined;
-  // 'primitive' : AES-GCM raw consistency test, no iter.
-  // 'selected'  : the iter the user picked in the Select.
-  // 'default'   : probed production default iter — always measured.
-  // 'both'      : selected === default (single row, treated as selected).
-  category: 'primitive' | 'selected' | 'default' | 'both';
-};
-
-type IAesGcmV2TableBackend =
-  | 'asmcrypto'
-  | 'fastNative'
-  | 'native'
-  | 'noble'
-  | 'quickNative'
-  | 'webcrypto';
-
-type IAesGcmV2BackendHeaderKind = 'aesGcm' | 'kdf';
-
-type IAesGcmV2DefaultProbe = {
-  aesGcmBackend?: string;
-  aesGcmOperation?: string;
-  configuredAesGcmBackend: string;
-  configuredIterations: number;
-  configuredPbkdf2Backend: string;
-  payloadIterations: number;
-  payloadVersion: string;
-  pbkdf2Backend?: string;
-  pbkdf2Iterations?: number;
-  aesGcmTableBackend: IAesGcmV2TableBackend | null;
-  pbkdf2TableBackend: IAesGcmV2TableBackend | null;
-  tableBackend: IAesGcmV2TableBackend | null;
-  time: number;
-};
-
-function getAesGcmV2TableBackendByPbkdf2Backend(
-  backend?: string,
-): IAesGcmV2TableBackend | null {
-  if (backend === 'asmcrypto') {
-    return 'asmcrypto';
-  }
-  if (backend === 'webcrypto') {
-    return 'webcrypto';
-  }
-  if (backend === 'react-native-quick-crypto') {
-    return 'quickNative';
-  }
-  if (backend === 'react-native-fast-pbkdf2') {
-    return 'fastNative';
-  }
-  if (backend === 'react-native-aes-crypto') {
-    return 'native';
-  }
-  if (backend === 'noble') {
-    return 'noble';
-  }
-  return null;
-}
-
-function getAesGcmV2TableBackendByAesGcmBackend(
-  backend?: string,
-): IAesGcmV2TableBackend | null {
-  if (backend === 'react-native-aes-crypto') {
-    return 'native';
-  }
-  if (backend === 'noble') {
-    return 'noble';
-  }
-  return null;
-}
-
-function getAesGcmV2BackendHeaderInfo({
-  backend,
-  kind,
-}: {
-  backend: IAesGcmV2TableBackend;
-  kind: IAesGcmV2BackendHeaderKind;
-}) {
-  if (kind === 'aesGcm') {
-    if (backend === 'native') {
-      return {
-        title: 'AES-GCM native',
-        message:
-          'react-native-aes-crypto: native AES-GCM encrypt/decrypt backend',
-      };
-    }
-    if (backend === 'noble') {
-      return {
-        title: 'AES-GCM noble',
-        message: '@noble/ciphers/aes: JS AES-GCM backend',
-      };
-    }
-    return {
-      title: 'AES-GCM backend',
-      message: 'This backend is not used by the AES-GCM raw table',
-    };
-  }
-  if (backend === 'asmcrypto') {
-    return {
-      title: 'PBKDF2 asm',
-      message: 'asmcrypto.js: Pbkdf2HmacSha256 implementation',
-    };
-  }
-  if (backend === 'webcrypto') {
-    return {
-      title: 'PBKDF2 webcrypto',
-      message: 'Web Crypto API: crypto.subtle PBKDF2 implementation',
-    };
-  }
-  if (backend === 'quickNative') {
-    return {
-      title: 'PBKDF2 quick',
-      message: 'react-native-quick-crypto: native PBKDF2 implementation',
-    };
-  }
-  if (backend === 'fastNative') {
-    return {
-      title: 'PBKDF2 fast',
-      message: 'react-native-fast-pbkdf2: native PBKDF2 implementation',
-    };
-  }
-  if (backend === 'native') {
-    return {
-      title: 'PBKDF2 native',
-      message: 'react-native-aes-crypto: native PBKDF2 implementation',
-    };
-  }
-  return {
-    title: 'PBKDF2 noble',
-    message: '@noble/hashes/pbkdf2: JS PBKDF2 implementation',
-  };
-}
-
-function classifyAesGcmV2TaskName(
-  name: string,
-  defaultIter: number,
-): {
-  op:
-    | 'AES-GCM'
-    | 'PBKDF2'
-    | 'encryptAsync'
-    | 'decryptAsync'
-    | 'probe'
-    | 'other';
-  iter: number | undefined;
-  backend: IAesGcmV2TableBackend | '—';
-  isProbe: boolean;
-} {
-  let op:
-    | 'AES-GCM'
-    | 'PBKDF2'
-    | 'encryptAsync'
-    | 'decryptAsync'
-    | 'probe'
-    | 'other' = 'other';
-  if (/^encryptAsync\b/.test(name)) op = 'encryptAsync';
-  else if (/^decryptAsync\b/.test(name)) op = 'decryptAsync';
-  else if (/^PBKDF2\b/.test(name)) op = 'PBKDF2';
-  else if (/^AES-GCM\b/.test(name)) op = 'AES-GCM';
-  else if (/^actual\b/.test(name)) op = 'probe';
-
-  const isProbe =
-    op === 'probe' ||
-    /^encryptAsync\s+\S+\s+actual/.test(name) ||
-    /^decryptAsync\s+\S+\s+actual/.test(name) ||
-    /^PBKDF2\s+default\(/.test(name) ||
-    /^encryptAsync\s+(default writes|default iterations|v2 prefix)/.test(
-      name,
-    ) ||
-    /^decryptAsync\s+reads v2 payload/.test(name) ||
-    /^actual\s+payload|^actual\s+PBKDF2|^actual\s+AES-GCM/.test(name);
-
-  let iter: number | undefined;
-  let backend: IAesGcmV2TableBackend | '—' = '—';
-  if (op === 'AES-GCM') {
-    iter = undefined;
-    backend = /^AES-GCM\s+noble\b/.test(name) ? 'noble' : 'native';
-  } else if (op === 'PBKDF2') {
-    const m = name.match(/\b(\d{4,})\b/);
-    if (m) iter = Number(m[1]);
-    if (/\basmcrypto\b/.test(name)) {
-      backend = 'asmcrypto';
-    } else if (/\bwebcrypto\b/.test(name)) {
-      backend = 'webcrypto';
-    } else if (/react-native-quick-crypto|\bquick-native\b/.test(name)) {
-      backend = 'quickNative';
-    } else if (/react-native-fast-pbkdf2|\bfast-native\b/.test(name)) {
-      backend = 'fastNative';
-    } else if (/\bnoble\b/.test(name)) backend = 'noble';
-    else if (/\bnative\b/.test(name) || /\bdefault\b/.test(name))
-      backend = 'native';
-  } else if (op === 'encryptAsync' || op === 'decryptAsync') {
-    const m = name.match(/\b(\d{4,})\b/);
-    if (m) iter = Number(m[1]);
-    if (/\basmcrypto\b/.test(name)) {
-      backend = 'asmcrypto';
-    } else if (/\bwebcrypto\b/.test(name)) {
-      backend = 'webcrypto';
-    } else if (/react-native-quick-crypto|\bquick-native\b/.test(name)) {
-      backend = 'quickNative';
-    } else if (/react-native-fast-pbkdf2|\bfast-native\b/.test(name)) {
-      backend = 'fastNative';
-    } else if (/\bnoble\b/.test(name)) backend = 'noble';
-    else if (/\bnative\b/.test(name)) backend = 'native';
-    else if (
-      /default writes|default iterations|reads v2 payload/.test(name) ||
-      /^encryptAsync\s+default\b/.test(name)
-    ) {
-      backend = 'native';
-      if (!iter) iter = defaultIter;
-    }
-  }
-  return { op, iter, backend, isProbe };
-}
-
-function buildAesGcmV2TableRows(payload: {
-  tasks: { name: string; time: number }[];
-  actualEncryptRuns: {
-    requestedIterations: number | 'default';
-    payloadIterations: number;
-    time: number;
-    pbkdf2Invocation?: { backend?: string };
-  }[];
-  defaultAesGcmBackend: IAesGcmV2TableBackend | null;
-  defaultBackend: IAesGcmV2TableBackend | null;
-  selectedIter: number;
-  defaultIter: number;
-  defaultIterForMarker: number | null;
-}): IAesGcmV2TableRow[] {
-  const pivot: Record<string, Record<string, Record<string, number>>> = {};
-  const fill = (
-    op: string,
-    iter: number | undefined,
-    backend: string,
-    time: number,
-  ) => {
-    const iterKey = iter === undefined ? '__no_iter' : String(iter);
-    pivot[iterKey] = pivot[iterKey] || {};
-    pivot[iterKey][op] = pivot[iterKey][op] || {};
-    if (pivot[iterKey][op][backend] === undefined) {
-      pivot[iterKey][op][backend] = time;
-    }
-  };
-  for (const t of payload.tasks) {
-    const c = classifyAesGcmV2TaskName(t.name, payload.defaultIter);
-    if (
-      !c.isProbe &&
-      ['AES-GCM', 'PBKDF2', 'encryptAsync', 'decryptAsync'].includes(c.op) &&
-      c.backend !== '—'
-    ) {
-      fill(c.op, c.iter, c.backend, t.time);
-    }
-  }
-  for (const run of payload.actualEncryptRuns) {
-    const iter =
-      run.requestedIterations === 'default'
-        ? payload.defaultIter
-        : Number(run.requestedIterations);
-    const backend = getAesGcmV2TableBackendByPbkdf2Backend(
-      run.pbkdf2Invocation?.backend,
-    );
-    if (backend) {
-      fill('encryptAsync', iter, backend, run.time);
-    }
-  }
-  const lookup = (
-    op: string,
-    iter: number | null,
-    backend: IAesGcmV2TableBackend,
-  ) => pivot[iter === null ? '__no_iter' : String(iter)]?.[op]?.[backend];
-
-  const selectedIsDefault = payload.selectedIter === payload.defaultIter;
-  const categorize = (
-    iter: number | null,
-  ): 'primitive' | 'selected' | 'default' | 'both' => {
-    if (iter === null) return 'primitive';
-    if (selectedIsDefault && iter === payload.defaultIter) return 'both';
-    if (iter === payload.selectedIter) return 'selected';
-    return 'default';
-  };
-  const rowDefs: { opName: string; op: string; iter: number | null }[] = [
-    { opName: 'AES-GCM', op: 'AES-GCM', iter: null },
-    { opName: 'pbkdf2', op: 'PBKDF2', iter: payload.selectedIter },
-    { opName: 'pbkdf2', op: 'PBKDF2', iter: payload.defaultIter },
-    { opName: 'encryptAsync', op: 'encryptAsync', iter: payload.selectedIter },
-    { opName: 'encryptAsync', op: 'encryptAsync', iter: payload.defaultIter },
-    { opName: 'decryptAsync', op: 'decryptAsync', iter: payload.selectedIter },
-    { opName: 'decryptAsync', op: 'decryptAsync', iter: payload.defaultIter },
-  ];
-  const seen = new Set<string>();
-  const rows: IAesGcmV2TableRow[] = [];
-  for (const r of rowDefs) {
-    const k = `${r.op}::${r.iter ?? 'null'}`;
-    if (!seen.has(k)) {
-      seen.add(k);
-      rows.push({
-        opName: r.opName,
-        actualDefaultAesGcmBackend: payload.defaultAesGcmBackend,
-        iter: r.iter,
-        actualDefaultBackend: payload.defaultBackend,
-        asmcrypto: lookup(r.op, r.iter, 'asmcrypto'),
-        fastNative: lookup(r.op, r.iter, 'fastNative'),
-        isActualDefaultIter:
-          payload.defaultIterForMarker !== null &&
-          r.iter === payload.defaultIterForMarker,
-        noble: lookup(r.op, r.iter, 'noble'),
-        native: lookup(r.op, r.iter, 'native'),
-        quickNative: lookup(r.op, r.iter, 'quickNative'),
-        webcrypto: lookup(r.op, r.iter, 'webcrypto'),
-        category: categorize(r.iter),
-      });
-    }
-  }
-  return rows;
-}
-
-function AESGcmV2Test() {
-  const [resultJson, setResultJson] = useState('');
-  const [tableRows, setTableRows] = useState<IAesGcmV2TableRow[]>([]);
-  const [actualOutputs, setActualOutputs] =
-    useState<IPbkdf2ActualOutputs | null>(null);
-  const [errorMessage, setErrorMessage] = useState('');
-  const [lastRunSelectedIter, setLastRunSelectedIter] = useState<number | null>(
-    null,
-  );
-  const [lastRunDefaultTableBackend, setLastRunDefaultTableBackend] =
-    useState<IAesGcmV2TableBackend | null>(null);
-  const [lastRunDefaultProbe, setLastRunDefaultProbe] =
-    useState<IAesGcmV2DefaultProbe | null>(null);
-  const [selectedIter, setSelectedIter] = useState<string>(
-    String(AES_GCM_V2_ITER_OPTIONS[0]),
-  );
-  const [running, setRunning] = useState(false);
-  const { copyText } = useClipboard();
-
-  const testAESGcmV2 = async (
-    iterationsToRun: number[] = [AES_GCM_V2_DEFAULT_ITER],
-  ) => {
-    try {
-      const { decryptAsync, decryptAsyncWithMetadata, encryptAsync } =
-        await loadCoreSecret();
-      const tasks: IRunAppCryptoTestTaskResult[] = [];
-      const data = Buffer.from('onekey-aes-gcm-v2-gallery-test', 'utf8');
-      const key = Buffer.from(
-        '000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f',
-        'hex',
-      );
-      const nonce = Buffer.from('202122232425262728292a2b', 'hex');
-      const aad = Buffer.from('onekey-gallery-aad-v1', 'utf8');
-      const password = 'onekey-gallery-password';
-      const v2MagicText = '1K_ENC_V2';
-      const v2MagicHex = Buffer.from(v2MagicText, 'utf8').toString('hex');
-      const legacyGcmMagicText = appCrypto.consts.AES_GCM_ENCRYPTION_MAGIC;
-      const legacyGcmMagicHex = Buffer.from(
-        legacyGcmMagicText,
-        'utf8',
-      ).toString('hex');
-      const configuredDefaultPbkdf2Backend =
-        appCrypto.pbkdf2.getPbkdf2BackendForCurrentPlatform();
-      const configuredDefaultAesGcmBackend =
-        appCrypto.aesGcm.getAesGcmBackendForCurrentPlatform();
-      const configuredDefaultIterations =
-        appCrypto.consts.PBKDF2_CURRENT_NUM_OF_ITERATIONS;
-      const actualEncryptRuns: Array<{
-        aesGcmInvocation: ReturnType<
-          typeof appCrypto.aesGcm.getLastAesGcmInvocation
-        >;
-        pbkdf2Invocation: ReturnType<
-          typeof appCrypto.pbkdf2.getLastPbkdf2Invocation
-        >;
-        payloadHeaderHex?: string;
-        payloadHeaderKind:
-          | 'legacy-cbc-no-magic-header'
-          | 'legacy-gcm'
-          | 'unknown'
-          | 'v2';
-        payloadHeaderText?: string;
-        payloadIterations: number;
-        payloadVersion: string;
-        requestedIterations: number | 'default';
-        time: number;
-      }> = [];
-      let shouldCooldownBeforeBenchmarkTask = false;
-      const waitBeforeBenchmarkTask = async () => {
-        if (!shouldCooldownBeforeBenchmarkTask) {
-          shouldCooldownBeforeBenchmarkTask = true;
-          return;
-        }
-        await waitForAesGcmV2BenchmarkCooldown();
-      };
-
-      const getKnownPayloadHeader = (encryptedHex: string) => {
-        if (encryptedHex.startsWith(v2MagicHex)) {
-          return {
-            payloadHeaderHex: v2MagicHex,
-            payloadHeaderKind: 'v2' as const,
-            payloadHeaderText: v2MagicText,
-          };
-        }
-        if (encryptedHex.startsWith(legacyGcmMagicHex)) {
-          return {
-            payloadHeaderHex: legacyGcmMagicHex,
-            payloadHeaderKind: 'legacy-gcm' as const,
-            payloadHeaderText: legacyGcmMagicText,
-          };
-        }
-        return {
-          payloadHeaderKind: 'legacy-cbc-no-magic-header' as const,
-        };
-      };
-
-      const encryptWithActualProbe = async (iterations?: number) => {
-        const debugCryptoProbeId = `crypto-gallery-${Date.now()}-${
-          actualEncryptRuns.length
-        }`;
-        appCrypto.pbkdf2.clearLastPbkdf2Invocation();
-        appCrypto.aesGcm.clearLastAesGcmInvocation();
-        appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(debugCryptoProbeId);
-        appCrypto.aesGcm.clearAesGcmInvocationByProbeId(debugCryptoProbeId);
-        const start = Date.now();
-        const encrypted = await encryptAsync({
-          password,
-          data,
-          allowRawPassword: true,
-          ...(iterations ? { iterations } : undefined),
-          debugCryptoProbeId,
-        });
-        const time = Date.now() - start;
-        const pbkdf2Invocation =
-          appCrypto.pbkdf2.getPbkdf2InvocationByProbeId(debugCryptoProbeId);
-        const aesGcmInvocation =
-          appCrypto.aesGcm.getAesGcmInvocationByProbeId(debugCryptoProbeId);
-        const encryptedHex = bufferUtils.bytesToHex(encrypted);
-        const payloadHeader = getKnownPayloadHeader(encryptedHex);
-        const metadata = await decryptAsyncWithMetadata({
-          password,
-          data: encrypted,
-          allowRawPassword: true,
-        });
-        const requestedIterations: number | 'default' = iterations ?? 'default';
-        const actualRun = {
-          aesGcmInvocation,
-          pbkdf2Invocation,
-          ...payloadHeader,
-          payloadIterations: metadata.iterations,
-          payloadVersion: metadata.version,
-          requestedIterations,
-          time,
-        };
-        actualEncryptRuns.push(actualRun);
-        appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(debugCryptoProbeId);
-        appCrypto.aesGcm.clearAesGcmInvocationByProbeId(debugCryptoProbeId);
-        return {
-          ...actualRun,
-          encrypted,
-          encryptedHex,
-          metadata,
-        };
-      };
-      await waitBeforeBenchmarkTask();
-      const encryptedV2 = await encryptWithActualProbe();
-      const actualDefaultIterations =
-        encryptedV2.pbkdf2Invocation?.iterations ??
-        encryptedV2.payloadIterations;
-      const probedDefaultIterations =
-        encryptedV2.pbkdf2Invocation?.iterations ?? null;
-      const actualDefaultPbkdf2Backend = encryptedV2.pbkdf2Invocation?.backend;
-      const actualDefaultAesGcmBackend = encryptedV2.aesGcmInvocation?.backend;
-      const actualDefaultTableBackend = getAesGcmV2TableBackendByPbkdf2Backend(
-        actualDefaultPbkdf2Backend,
-      );
-      const actualDefaultAesGcmTableBackend =
-        getAesGcmV2TableBackendByAesGcmBackend(actualDefaultAesGcmBackend);
-      const actualDefaultProbe: IAesGcmV2DefaultProbe = {
-        aesGcmBackend: actualDefaultAesGcmBackend,
-        aesGcmOperation: encryptedV2.aesGcmInvocation?.operation,
-        aesGcmTableBackend: actualDefaultAesGcmTableBackend,
-        configuredAesGcmBackend: configuredDefaultAesGcmBackend,
-        configuredIterations: configuredDefaultIterations,
-        configuredPbkdf2Backend: configuredDefaultPbkdf2Backend,
-        payloadIterations: encryptedV2.payloadIterations,
-        payloadVersion: encryptedV2.payloadVersion,
-        pbkdf2Backend: actualDefaultPbkdf2Backend,
-        pbkdf2Iterations: encryptedV2.pbkdf2Invocation?.iterations,
-        pbkdf2TableBackend: actualDefaultTableBackend,
-        tableBackend: actualDefaultTableBackend,
-        time: encryptedV2.time,
-      };
-
-      const nobleEncrypted = appCrypto.aesGcm.aesGcmEncryptByNoble({
-        nonce,
-        key,
-        data,
-        aad,
-      });
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: bufferUtils.bytesToHex(nobleEncrypted),
-          name: 'AES-GCM default wrapper matches noble',
-          fn: () =>
-            appCrypto.aesGcm.aesGcmEncrypt({
-              nonce,
-              key,
-              data,
-              aad,
-            }),
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: bufferUtils.bytesToHex(nobleEncrypted),
-          name: 'AES-GCM native wrapper matches noble',
-          fn: () =>
-            appCrypto.aesGcm.aesGcmEncryptByRNAes({
-              nonce,
-              key,
-              data,
-              aad,
-            }),
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: bufferUtils.bytesToHex(data),
-          name: 'AES-GCM native decrypt',
-          fn: () =>
-            appCrypto.aesGcm.aesGcmDecryptByRNAes({
-              nonce,
-              key,
-              data: nobleEncrypted,
-              aad,
-            }),
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: bufferUtils.bytesToHex(nobleEncrypted),
-          name: 'AES-GCM noble encrypt',
-          fn: () =>
-            appCrypto.aesGcm.aesGcmEncryptByNoble({
-              nonce,
-              key,
-              data,
-              aad,
-            }),
-        }),
-      );
-
-      // Boundary vectors aligned with upstream PR
-      // https://github.com/OneKeyHQ/app-modules/pull/55 Codex review +
-      // strict empty-string guard rolled out in commit 75031eb: empty
-      // plaintext must be rejected at the JS invoke check, native must
-      // reject empty AAD, and both backends must reject truncated /
-      // tampered ciphertexts. These are the cases jest-on-node alone
-      // cannot fully exercise because they cross into native CryptoKit
-      // and Android Cipher.
-      const emptyData = Buffer.alloc(0);
-      const emptyAad = Buffer.alloc(0);
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM noble rejects empty plaintext',
-          fn: () => {
-            try {
-              appCrypto.aesGcm.aesGcmEncryptByNoble({
-                nonce,
-                key,
-                data: emptyData,
-                aad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM native rejects empty plaintext',
-          fn: async () => {
-            try {
-              await appCrypto.aesGcm.aesGcmEncryptByRNAes({
-                nonce,
-                key,
-                data: emptyData,
-                aad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      // Empty AAD is rejected at the JS invoke check so noble and native
-      // stay in lockstep across platforms (otherwise desktop/web on noble
-      // would silently accept an empty AAD that mobile native rejects).
-      // Both backends must surface the same REJECTED result.
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM noble rejects empty AAD',
-          fn: () => {
-            try {
-              appCrypto.aesGcm.aesGcmEncryptByNoble({
-                nonce,
-                key,
-                data,
-                aad: emptyAad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM native rejects empty AAD',
-          fn: async () => {
-            try {
-              await appCrypto.aesGcm.aesGcmEncryptByRNAes({
-                nonce,
-                key,
-                data,
-                aad: emptyAad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      // 8-byte ciphertext is shorter than the 16-byte GCM tag. Both
-      // backends must reject; otherwise tag verification could be skipped.
-      const shortCiphertext = Buffer.from('0011223344556677', 'hex');
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM native rejects ciphertext shorter than tag',
-          fn: async () => {
-            try {
-              await appCrypto.aesGcm.aesGcmDecryptByRNAes({
-                nonce,
-                key,
-                data: shortCiphertext,
-                aad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM noble rejects ciphertext shorter than tag',
-          fn: () => {
-            try {
-              appCrypto.aesGcm.aesGcmDecryptByNoble({
-                nonce,
-                key,
-                data: shortCiphertext,
-                aad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      // Tampered AAD must fail tag verification, even when the ciphertext
-      // bytes are otherwise valid.
-      const wrongAad = Buffer.from('onekey-gallery-aad-WRONG', 'utf8');
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM native rejects decrypt with wrong AAD',
-          fn: async () => {
-            try {
-              await appCrypto.aesGcm.aesGcmDecryptByRNAes({
-                nonce,
-                key,
-                data: nobleEncrypted,
-                aad: wrongAad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'REJECTED',
-          name: 'AES-GCM noble rejects decrypt with wrong AAD',
-          fn: () => {
-            try {
-              appCrypto.aesGcm.aesGcmDecryptByNoble({
-                nonce,
-                key,
-                data: nobleEncrypted,
-                aad: wrongAad,
-              });
-              return 'NOT_REJECTED';
-            } catch {
-              return 'REJECTED';
-            }
-          },
-        }),
-      );
-
-      const hashedPassword = await appCrypto.hash.sha256(
-        Buffer.from(password, 'utf8'),
-      );
-      const salt = Buffer.from(
-        '303132333435363738393a3b3c3d3e3f404142434445464748494a4b4c4d4e4f',
-        'hex',
-      );
-
-      // Always include the default iteration count observed by the encrypt
-      // probe, so the table follows the path that actually ran instead of a
-      // configured constant. Noble at high iteration counts is gated below.
-      const NOBLE_KDF_MAX_ITER = 10_000;
-      const userSelectedIter = iterationsToRun[0] ?? actualDefaultIterations;
-      const effectiveIterations = Array.from(
-        new Set([...iterationsToRun, actualDefaultIterations]),
-      ).toSorted((a, b) => a - b);
-
-      const capturedOutputs: IPbkdf2ActualOutputs = {
-        iter: userSelectedIter,
-        sha256: {},
-        sha512: {},
-      };
-      const captureIfSelectedIter = (
-        bucket: 'sha256' | 'sha512',
-        backend:
-          | 'defaultBackend'
-          | 'fastNative'
-          | 'native'
-          | 'noble'
-          | 'quickNative',
-        iter: number,
-        hex: string | undefined,
-      ) => {
-        if (iter !== userSelectedIter || !hex) return;
-        const target = capturedOutputs[bucket] as Record<
-          string,
-          string | undefined
-        >;
-        target[backend] = Buffer.from(hex, 'hex').toString('base64');
-      };
-      const hashedPasswordHex = bufferUtils.bytesToHex(hashedPassword);
-      const saltHex = bufferUtils.bytesToHex(salt);
-
-      for (const iterations of effectiveIterations) {
-        await waitBeforeBenchmarkTask();
-        const defaultKey = await appCrypto.pbkdf2.pbkdf2({
-          password: hashedPassword,
-          salt,
-          iterations,
-        });
-
-        // Pin every backend at this iter to a canonical third-party hex
-        // (Node crypto / noble pre-computed offline). Falls back to
-        // intra-device consistency if a new iter is added without updating
-        // AES_GCM_V2_PBKDF2_EXPECTED.
-        const canonicalHex: string | undefined =
-          AES_GCM_V2_PBKDF2_EXPECTED[iterations];
-        const expectedHex = canonicalHex ?? bufferUtils.bytesToHex(defaultKey);
-        captureIfSelectedIter(
-          'sha256',
-          'defaultBackend',
-          iterations,
-          bufferUtils.bytesToHex(defaultKey),
-        );
-
-        // Marked `actual ...` so the classifier treats it as a probe and
-        // excludes it from the timing pivot. It is a pure correctness check
-        // that re-hashes a pre-computed buffer — its time would otherwise
-        // pollute the native column (first-write-wins in pivot).
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: expectedHex,
-            name: `actual PBKDF2 canonical default(${actualDefaultPbkdf2Backend ?? 'missing'}) ${iterations}`,
-            fn: () => bufferUtils.bytesToHex(defaultKey),
-          }),
-        );
-
-        await waitBeforeBenchmarkTask();
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: expectedHex,
-            name: `PBKDF2 default(${actualDefaultPbkdf2Backend ?? 'missing'}) ${iterations}`,
-            fn: () =>
-              appCrypto.pbkdf2.pbkdf2({
-                password: hashedPassword,
-                salt,
-                iterations,
-              }),
-          }),
-        );
-
-        await waitBeforeBenchmarkTask();
-        const nativeTask = await runAppCryptoTestTask({
-          expect: expectedHex,
-          name: `PBKDF2 native ${iterations}`,
-          fn: () =>
-            appCrypto.pbkdf2.pbkdf2ByRNAes({
-              password: hashedPassword,
-              salt,
-              iterations,
-            }),
-        });
-        tasks.push(nativeTask);
-        captureIfSelectedIter(
-          'sha256',
-          'native',
-          iterations,
-          nativeTask.result,
-        );
-
-        if (platformEnv.isNative) {
-          await waitBeforeBenchmarkTask();
-          const quickTask = await runAppCryptoTestTask({
-            expect: expectedHex,
-            name: `PBKDF2 quick-native ${iterations}`,
-            fn: () =>
-              appCrypto.pbkdf2.pbkdf2ByRNQuickCrypto({
-                password: hashedPassword,
-                salt,
-                iterations,
-              }),
-          });
-          tasks.push(quickTask);
-          captureIfSelectedIter(
-            'sha256',
-            'quickNative',
-            iterations,
-            quickTask.result,
-          );
-
-          await waitBeforeBenchmarkTask();
-          const fastTask = await runAppCryptoTestTask({
-            expect: expectedHex,
-            name: `PBKDF2 fast-native ${iterations}`,
-            fn: () =>
-              appCrypto.pbkdf2.pbkdf2ByRNFastPbkdf2({
-                password: hashedPassword,
-                salt,
-                iterations,
-              }),
-          });
-          tasks.push(fastTask);
-          captureIfSelectedIter(
-            'sha256',
-            'fastNative',
-            iterations,
-            fastTask.result,
-          );
-        }
-
-        if (iterations <= NOBLE_KDF_MAX_ITER) {
-          await waitBeforeBenchmarkTask();
-          const nobleTask = await runAppCryptoTestTask({
-            expect: expectedHex,
-            name: `PBKDF2 noble ${iterations}`,
-            fn: () =>
-              appCrypto.pbkdf2.pbkdf2ByNoble({
-                password: hashedPassword,
-                salt,
-                iterations,
-              }),
-          });
-          tasks.push(nobleTask);
-          captureIfSelectedIter(
-            'sha256',
-            'noble',
-            iterations,
-            nobleTask.result,
-          );
-        }
-
-        // SHA-512 backends (TON / BIP39 path protection). All names are
-        // `actual ...` so they bypass the timing pivot. Correctness is still
-        // recorded via tasks[], and selected-iter captures feed the actual
-        // outputs table. Skip when no canonical vector is pinned (defensive:
-        // future iter additions should not silently break the assertion).
-        const sha512Canonical: string | undefined =
-          AES_GCM_V2_PBKDF2_SHA512_EXPECTED[iterations];
-        if (sha512Canonical) {
-          await waitBeforeBenchmarkTask();
-          const nativeSha512Task = await runAppCryptoTestTask({
-            expect: sha512Canonical,
-            name: `actual PBKDF2-SHA512 native ${iterations}`,
-            fn: () =>
-              pbkdf2Sha512ByRNAes(hashedPasswordHex, saltHex, iterations),
-          });
-          tasks.push(nativeSha512Task);
-          captureIfSelectedIter(
-            'sha512',
-            'native',
-            iterations,
-            nativeSha512Task.result,
-          );
-
-          if (platformEnv.isNative) {
-            await waitBeforeBenchmarkTask();
-            const quickSha512Task = await runAppCryptoTestTask({
-              expect: sha512Canonical,
-              name: `actual PBKDF2-SHA512 quick-native ${iterations}`,
-              fn: () =>
-                pbkdf2Sha512ByRNQuickCrypto(hashedPassword, salt, iterations),
-            });
-            tasks.push(quickSha512Task);
-            captureIfSelectedIter(
-              'sha512',
-              'quickNative',
-              iterations,
-              quickSha512Task.result,
-            );
-
-            await waitBeforeBenchmarkTask();
-            const fastSha512Task = await runAppCryptoTestTask({
-              expect: sha512Canonical,
-              name: `actual PBKDF2-SHA512 fast-native ${iterations}`,
-              fn: () => pbkdf2Sha512ByRNFast(hashedPassword, salt, iterations),
-            });
-            tasks.push(fastSha512Task);
-            captureIfSelectedIter(
-              'sha512',
-              'fastNative',
-              iterations,
-              fastSha512Task.result,
-            );
-          }
-
-          if (iterations <= NOBLE_KDF_MAX_ITER) {
-            await waitBeforeBenchmarkTask();
-            const nobleSha512Task = await runAppCryptoTestTask({
-              expect: sha512Canonical,
-              name: `actual PBKDF2-SHA512 noble ${iterations}`,
-              fn: () => pbkdf2Sha512ByNoble(hashedPassword, salt, iterations),
-            });
-            tasks.push(nobleSha512Task);
-            captureIfSelectedIter(
-              'sha512',
-              'noble',
-              iterations,
-              nobleSha512Task.result,
-            );
-          }
-        }
-
-        await waitBeforeBenchmarkTask();
-        const encryptedByIterations = await encryptWithActualProbe(iterations);
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: 'true',
-            name: `encryptAsync v2 prefix ${iterations}`,
-            fn: () =>
-              String(encryptedByIterations.encryptedHex.startsWith(v2MagicHex)),
-          }),
-        );
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: String(iterations),
-            name: `actual payload iterations ${iterations}`,
-            fn: () => String(encryptedByIterations.metadata.iterations),
-          }),
-        );
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: String(iterations),
-            name: `actual PBKDF2 probe iterations ${iterations}`,
-            fn: () =>
-              String(
-                encryptedByIterations.pbkdf2Invocation?.iterations ?? 'missing',
-              ),
-          }),
-        );
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect:
-              actualDefaultPbkdf2Backend ?? '__missing_default_pbkdf2_probe__',
-            name: `actual PBKDF2 probe backend ${iterations}`,
-            fn: () =>
-              encryptedByIterations.pbkdf2Invocation?.backend ?? 'missing',
-          }),
-        );
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect:
-              actualDefaultAesGcmBackend ?? '__missing_default_aes_gcm_probe__',
-            name: `actual AES-GCM probe backend ${iterations}`,
-            fn: () =>
-              encryptedByIterations.aesGcmInvocation?.backend ?? 'missing',
-          }),
-        );
-
-        tasks.push(
-          await runAppCryptoTestTask({
-            expect: 'encrypt',
-            name: `actual AES-GCM probe operation ${iterations}`,
-            fn: () =>
-              encryptedByIterations.aesGcmInvocation?.operation ?? 'missing',
-          }),
-        );
-
-        // Backend matrix: cover encryptAsync + decryptAsync for every iter
-        // that appears in the table. Native/fast-native use concrete PBKDF2
-        // backends so the DevSettings fast switch cannot alter these timings.
-        // The native path runs at any iter; the noble path is gated by
-        // NOBLE_KDF_MAX_ITER because noble PBKDF2 above ~10k iterations
-        // freezes the device for tens of seconds.
-        const backendConfigs = [
-          {
-            enabled: platformEnv.isNative,
-            expectedGcmBackend: 'react-native-aes-crypto',
-            expectedPbkdf2Backend: 'react-native-quick-crypto',
-            gcmBackend: 'native',
-            kdfBackend: 'react-native-quick-crypto',
-            label: 'quick-native',
-          },
-          {
-            enabled: true,
-            expectedGcmBackend: 'react-native-aes-crypto',
-            expectedPbkdf2Backend: 'react-native-aes-crypto',
-            gcmBackend: 'native',
-            kdfBackend: 'react-native-aes-crypto',
-            label: 'native',
-          },
-          {
-            enabled: platformEnv.isNative,
-            expectedGcmBackend: 'react-native-aes-crypto',
-            expectedPbkdf2Backend: 'react-native-fast-pbkdf2',
-            gcmBackend: 'native',
-            kdfBackend: 'react-native-fast-pbkdf2',
-            label: 'fast-native',
-          },
-          {
-            enabled: iterations <= NOBLE_KDF_MAX_ITER,
-            expectedGcmBackend: 'noble',
-            expectedPbkdf2Backend: 'noble',
-            gcmBackend: 'noble',
-            kdfBackend: 'noble',
-            label: 'noble',
-          },
-        ] as const;
-        for (const backendConfig of backendConfigs.filter((b) => b.enabled)) {
-          const encProbeId = `crypto-gallery-mx-${backendConfig.label}-enc-${iterations}-${Date.now()}-${tasks.length}`;
-          appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(encProbeId);
-          appCrypto.aesGcm.clearAesGcmInvocationByProbeId(encProbeId);
-          let encryptedByBackend: Buffer | undefined;
-          await waitBeforeBenchmarkTask();
-          tasks.push(
-            await runAppCryptoTestTask({
-              expect: 'true',
-              name: `encryptAsync ${backendConfig.label} ${iterations}`,
-              fn: async () => {
-                encryptedByBackend = await encryptAsync({
-                  password,
-                  data,
-                  allowRawPassword: true,
-                  iterations,
-                  debugCryptoProbeId: encProbeId,
-                  kdfBackend: backendConfig.kdfBackend,
-                  gcmBackend: backendConfig.gcmBackend,
-                });
-                return String(
-                  bufferUtils
-                    .bytesToHex(encryptedByBackend)
-                    .startsWith(v2MagicHex),
-                );
-              },
-            }),
-          );
-          const encPbkdf2 =
-            appCrypto.pbkdf2.getPbkdf2InvocationByProbeId(encProbeId);
-          const encAesGcm =
-            appCrypto.aesGcm.getAesGcmInvocationByProbeId(encProbeId);
-          tasks.push(
-            await runAppCryptoTestTask({
-              expect: backendConfig.expectedPbkdf2Backend,
-              name: `encryptAsync ${backendConfig.label} actual pbkdf2 backend ${iterations}`,
-              fn: () => encPbkdf2?.backend ?? 'missing',
-            }),
-          );
-          tasks.push(
-            await runAppCryptoTestTask({
-              expect: backendConfig.expectedGcmBackend,
-              name: `encryptAsync ${backendConfig.label} actual aesGcm backend ${iterations}`,
-              fn: () => encAesGcm?.backend ?? 'missing',
-            }),
-          );
-          appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(encProbeId);
-          appCrypto.aesGcm.clearAesGcmInvocationByProbeId(encProbeId);
-
-          if (encryptedByBackend) {
-            const decProbeId = `crypto-gallery-mx-${backendConfig.label}-dec-${iterations}-${Date.now()}-${tasks.length}`;
-            appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(decProbeId);
-            appCrypto.aesGcm.clearAesGcmInvocationByProbeId(decProbeId);
-            await waitBeforeBenchmarkTask();
-            tasks.push(
-              await runAppCryptoTestTask({
-                expect: bufferUtils.bytesToHex(data),
-                name: `decryptAsync ${backendConfig.label} ${iterations}`,
-                fn: () =>
-                  decryptAsync({
-                    password,
-                    data: encryptedByBackend!,
-                    allowRawPassword: true,
-                    debugCryptoProbeId: decProbeId,
-                    kdfBackend: backendConfig.kdfBackend,
-                    gcmBackend: backendConfig.gcmBackend,
-                  }),
-              }),
-            );
-            const decPbkdf2 =
-              appCrypto.pbkdf2.getPbkdf2InvocationByProbeId(decProbeId);
-            const decAesGcm =
-              appCrypto.aesGcm.getAesGcmInvocationByProbeId(decProbeId);
-            tasks.push(
-              await runAppCryptoTestTask({
-                expect: backendConfig.expectedPbkdf2Backend,
-                name: `decryptAsync ${backendConfig.label} actual pbkdf2 backend ${iterations}`,
-                fn: () => decPbkdf2?.backend ?? 'missing',
-              }),
-            );
-            tasks.push(
-              await runAppCryptoTestTask({
-                expect: backendConfig.expectedGcmBackend,
-                name: `decryptAsync ${backendConfig.label} actual aesGcm backend ${iterations}`,
-                fn: () => decAesGcm?.backend ?? 'missing',
-              }),
-            );
-            appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(decProbeId);
-            appCrypto.aesGcm.clearAesGcmInvocationByProbeId(decProbeId);
-          }
-        }
-      }
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'true',
-          name: 'encryptAsync default writes 1K_ENC_V2',
-          fn: () => String(encryptedV2.encryptedHex.startsWith(v2MagicHex)),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: String(encryptedV2.payloadIterations),
-          name: 'encryptAsync default iterations',
-          fn: () => String(encryptedV2.metadata.iterations),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: String(encryptedV2.payloadIterations),
-          name: 'actual default PBKDF2 probe iterations',
-          fn: () =>
-            String(encryptedV2.pbkdf2Invocation?.iterations ?? 'missing'),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            actualDefaultPbkdf2Backend ?? '__missing_default_pbkdf2_probe__',
-          name: 'actual default PBKDF2 probe backend',
-          fn: () => encryptedV2.pbkdf2Invocation?.backend ?? 'missing',
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            actualDefaultAesGcmBackend ?? '__missing_default_aes_gcm_probe__',
-          name: 'actual default AES-GCM probe backend',
-          fn: () => encryptedV2.aesGcmInvocation?.backend ?? 'missing',
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'encrypt',
-          name: 'actual default AES-GCM probe operation',
-          fn: () => encryptedV2.aesGcmInvocation?.operation ?? 'missing',
-        }),
-      );
-
-      await waitBeforeBenchmarkTask();
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: bufferUtils.bytesToHex(data),
-          name: 'decryptAsync reads v2 payload',
-          fn: () =>
-            decryptAsync({
-              password,
-              data: encryptedV2.encrypted,
-              allowRawPassword: true,
-            }),
-        }),
-      );
-
-      const nextTableRows = buildAesGcmV2TableRows({
-        tasks: tasks.map((t) => ({ name: t.name, time: t.time })),
-        actualEncryptRuns,
-        defaultAesGcmBackend: actualDefaultAesGcmTableBackend,
-        defaultBackend: actualDefaultTableBackend,
-        selectedIter: userSelectedIter,
-        defaultIter: actualDefaultIterations,
-        defaultIterForMarker: probedDefaultIterations,
-      });
-      setResultJson(
-        stringifyCryptoGalleryTablePayload({
-          platform: {
-            isNative: platformEnv.isNative,
-            isNativeIOS: platformEnv.isNativeIOS,
-            isNativeAndroid: platformEnv.isNativeAndroid,
-          },
-          actualEncryptRuns,
-          defaultPath: {
-            pbkdf2: actualDefaultPbkdf2Backend,
-            aesGcm: actualDefaultAesGcmBackend,
-            iterations: actualDefaultIterations,
-            actualProbe: actualDefaultProbe,
-            configured: {
-              pbkdf2: configuredDefaultPbkdf2Backend,
-              aesGcm: configuredDefaultAesGcmBackend,
-              iterations: configuredDefaultIterations,
-            },
-          },
-          v2MagicHex,
-          v2MagicText,
-          legacyGcmMagicHex,
-          legacyGcmMagicText,
-          encryptedV2Header: getKnownPayloadHeader(encryptedV2.encryptedHex),
-          tableRows: nextTableRows,
-          actualOutputs: capturedOutputs,
-          tasks,
-        }),
-      );
-      setLastRunSelectedIter(userSelectedIter);
-      setLastRunDefaultProbe(actualDefaultProbe);
-      setLastRunDefaultTableBackend(actualDefaultTableBackend);
-      setTableRows(nextTableRows);
-      setActualOutputs(capturedOutputs);
-      setErrorMessage('');
-
-      const allPassed = tasks.every(
-        (t) => t.isCorrect === AppCryptoTestEmoji.isCorrect,
-      );
-      if (allPassed) {
-        Toast.success({
-          title: 'AES-GCM PBKDF2 (v2 test) passed',
-        });
-      } else {
-        Toast.error({
-          title: 'AES-GCM PBKDF2 (v2 test) failed',
-        });
-      }
-    } catch (error) {
-      setErrorMessage((error as Error).message);
-      setResultJson('');
-      setTableRows([]);
-      setActualOutputs(null);
-      setLastRunDefaultTableBackend(null);
-      setLastRunDefaultProbe(null);
-      Toast.error({
-        title: `AES-GCM PBKDF2 (v2 test) failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  const selectedIterNumber = Number(selectedIter);
-  const fmtCell = (v: number | undefined) =>
-    v === undefined ? '-' : `${v} ms`;
-  // Tiered color for ms values:
-  //   undefined         → muted '—'
-  //   ≤ 10 ms           → success (green)
-  //   < 100 ms          → default text
-  //   100 .. < 500 ms   → caution (yellow/orange)
-  //   ≥ 500 ms          → critical (red)
-  const msColor = (v: number | undefined): string => {
-    if (v === undefined) return '$textDisabled';
-    if (v <= 10) return '$textSuccess';
-    if (v < 100) return '$text';
-    if (v < 500) return '$textCaution';
-    return '$textCritical';
-  };
-  const formatBase64AsShortHex = (value: string) => {
-    const hex = Buffer.from(value, 'base64').toString('hex');
-    return `${hex.slice(0, 8)}...${hex.slice(-8)}`;
-  };
-  const backendColumns: {
-    backend: IAesGcmV2TableBackend;
-    title: string;
-  }[] = [
-    { backend: 'asmcrypto', title: 'asm' },
-    { backend: 'webcrypto', title: 'webcrypto' },
-    { backend: 'noble', title: 'noble' },
-    { backend: 'quickNative', title: 'quick' },
-    { backend: 'fastNative', title: 'fast' },
-    { backend: 'native', title: 'native' },
-  ];
-  const aesGcmTableRows = tableRows.filter((row) => row.iter === null);
-  const kdfTimingTableRows = tableRows.filter((row) => row.iter !== null);
-  const aesGcmBackendColumns = backendColumns.filter(({ backend }) => {
-    if (backend !== 'noble' && backend !== 'native') {
-      return false;
-    }
-    if (lastRunDefaultProbe?.aesGcmTableBackend === backend) {
-      return true;
-    }
-    if (aesGcmTableRows.some((row) => row[backend] !== undefined)) {
-      return true;
-    }
-    if (backend === 'noble') {
-      return true;
-    }
-    return platformEnv.isNative && backend === 'native';
-  });
-  const kdfBackendColumns = backendColumns.filter(({ backend }) => {
-    if (lastRunDefaultTableBackend === backend) {
-      return true;
-    }
-    if (kdfTimingTableRows.some((row) => row[backend] !== undefined)) {
-      return true;
-    }
-    if (platformEnv.isNative) {
-      return (
-        backend === 'quickNative' ||
-        backend === 'fastNative' ||
-        backend === 'native'
-      );
-    }
-    return backend === 'asmcrypto' || backend === 'webcrypto';
-  });
-  const operationColumnFlexBasis = '28%';
-  const aesGcmOperationColumnFlexBasis = '44%';
-  const aesGcmBackendColumnFlexBasis: `${number}%` = `${
-    56 / Math.max(aesGcmBackendColumns.length, 1)
-  }%`;
-  const kdfBackendColumnFlexBasis: `${number}%` = `${
-    72 / Math.max(kdfBackendColumns.length, 1)
-  }%`;
-  const hasCompleteDefaultProbe = Boolean(
-    lastRunDefaultProbe?.pbkdf2Backend &&
-    lastRunDefaultProbe.pbkdf2Iterations !== undefined &&
-    lastRunDefaultProbe.aesGcmBackend,
-  );
-  const renderBackendHeader = (
-    backend: IAesGcmV2TableBackend,
-    defaultBackend: IAesGcmV2TableBackend | null | undefined,
-    kind: IAesGcmV2BackendHeaderKind,
-    title: string,
-  ) => (
-    <YStack
-      alignItems="flex-end"
-      borderBottomColor="$textInteractive"
-      borderBottomWidth={1}
-      borderStyle="dashed"
-      cursor="pointer"
-      hitSlop={NATIVE_HIT_SLOP}
-      hoverStyle={{ opacity: 0.8 }}
-      m={-4}
-      onPress={() => {
-        Toast.message(getAesGcmV2BackendHeaderInfo({ backend, kind }));
-      }}
-      p="$1"
-      pressStyle={{ opacity: 0.7 }}
-    >
-      <XStack gap="$1" alignItems="center">
-        {defaultBackend === backend ? (
-          <SizableText size="$bodySmMedium">
-            {CRYPTO_GALLERY_DEFAULT_PATH_EMOJI}
-          </SizableText>
-        ) : null}
-        <SizableText size="$bodySmMedium" color="$textInteractive">
-          {title}
-        </SizableText>
-      </XStack>
-    </YStack>
-  );
-  const renderBackendCell = (
-    row: IAesGcmV2TableRow,
-    backend: IAesGcmV2TableBackend,
-    value: number | undefined,
-  ) => (
-    <XStack gap="$1" alignItems="center" justifyContent="flex-end">
-      {(row.iter === null && row.actualDefaultAesGcmBackend === backend) ||
-      (row.isActualDefaultIter && row.actualDefaultBackend === backend) ? (
-        <SizableText size="$bodyMd">
-          {CRYPTO_GALLERY_DEFAULT_PATH_EMOJI}
-        </SizableText>
-      ) : null}
-      <SizableText size="$bodyMd" color={msColor(value)}>
-        {fmtCell(value)}
-      </SizableText>
-    </XStack>
-  );
-  return (
-    <PartContainer title="AES-GCM PBKDF2 (v2 test)">
-      <XStack gap="$3" alignItems="center" flexWrap="wrap">
-        <Select
-          items={AES_GCM_V2_ITER_OPTIONS.map((iter) => ({
-            value: String(iter),
-            label: `${iter.toLocaleString()} iterations`,
-          }))}
-          value={selectedIter}
-          onChange={setSelectedIter}
-          title="PBKDF2 iterations"
-          renderTrigger={({ label, onPress, disabled }) => (
-            <Button onPress={onPress} disabled={disabled}>
-              {label || `${selectedIterNumber.toLocaleString()} iterations`}
-            </Button>
-          )}
-        />
-        <Button
-          variant="primary"
-          loading={running}
-          disabled={running}
-          onPress={async () => {
-            setRunning(true);
-            try {
-              await runCryptoGalleryTestExclusive(() =>
-                testAESGcmV2([selectedIterNumber]),
-              );
-            } finally {
-              setRunning(false);
-            }
-          }}
-        >
-          Run Test
-        </Button>
-        {resultJson ? (
-          <Button
-            size="small"
-            onPress={() => {
-              copyText(resultJson);
-            }}
-          >
-            Copy raw JSON
-          </Button>
-        ) : null}
-      </XStack>
-      <SizableText size="$bodySm" color="$textSubdued">
-        Will run iter={selectedIterNumber.toLocaleString()} and the actual
-        default encrypt path detected by probe. Quick/fast/native timings force
-        concrete backends and ignore the DevSettings fast switch. Noble is
-        skipped for iter &gt; 10,000. Benchmark tasks wait{' '}
-        {AES_GCM_V2_BENCHMARK_COOLDOWN_MS}ms between runs. Quick/fast are
-        PBKDF2-only; AES-GCM raw uses noble/native.
-      </SizableText>
-
-      {lastRunDefaultProbe ? (
-        <SizableText size="$bodySm" color="$textSubdued">
-          {hasCompleteDefaultProbe ? CRYPTO_GALLERY_DEFAULT_PATH_EMOJI : ''}
-          {hasCompleteDefaultProbe ? ' ' : ''}Actual default probe: PBKDF2{' '}
-          {lastRunDefaultProbe.pbkdf2Backend ?? '-'} @{' '}
-          {(
-            lastRunDefaultProbe.pbkdf2Iterations ??
-            lastRunDefaultProbe.payloadIterations
-          ).toLocaleString()}{' '}
-          iterations, AES-GCM {lastRunDefaultProbe.aesGcmBackend ?? '-'}.
-        </SizableText>
-      ) : null}
-
-      {aesGcmTableRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <Stack flexBasis={aesGcmOperationColumnFlexBasis}>
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                AES-GCM raw
-              </SizableText>
-            </Stack>
-            {aesGcmBackendColumns.map(({ backend, title }) => (
-              <Stack
-                key={backend}
-                flexBasis={aesGcmBackendColumnFlexBasis}
-                alignItems="flex-end"
-              >
-                {renderBackendHeader(
-                  backend,
-                  lastRunDefaultProbe?.aesGcmTableBackend,
-                  'aesGcm',
-                  title,
-                )}
-              </Stack>
-            ))}
-          </XStack>
-          {aesGcmTableRows.map((row, idx) => (
-            <XStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={idx === aesGcmTableRows.length - 1 ? 0 : 1}
-              borderBottomColor="$borderSubdued"
-              alignItems="center"
-            >
-              <Stack flexBasis={aesGcmOperationColumnFlexBasis}>
-                <SizableText size="$bodyMd">{row.opName}</SizableText>
-              </Stack>
-              {aesGcmBackendColumns.map(({ backend }) => (
-                <Stack
-                  key={backend}
-                  flexBasis={aesGcmBackendColumnFlexBasis}
-                  alignItems="flex-end"
-                >
-                  {renderBackendCell(row, backend, row[backend])}
-                </Stack>
-              ))}
-            </XStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload({
-                    defaultProbe: lastRunDefaultProbe,
-                    rows: aesGcmTableRows,
-                  }),
-                );
-              }}
-            >
-              Copy AES-GCM table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {kdfTimingTableRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <Stack flexBasis={operationColumnFlexBasis}>
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                PBKDF2 / wrapper timing
-              </SizableText>
-            </Stack>
-            {kdfBackendColumns.map(({ backend, title }) => (
-              <Stack
-                key={backend}
-                flexBasis={kdfBackendColumnFlexBasis}
-                alignItems="flex-end"
-              >
-                {renderBackendHeader(
-                  backend,
-                  lastRunDefaultTableBackend,
-                  'kdf',
-                  title,
-                )}
-              </Stack>
-            ))}
-          </XStack>
-          {kdfTimingTableRows.map((row, idx) => (
-            <XStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={idx === kdfTimingTableRows.length - 1 ? 0 : 1}
-              borderBottomColor="$borderSubdued"
-              alignItems="center"
-            >
-              <Stack flexBasis={operationColumnFlexBasis}>
-                <SizableText size="$bodyMd">{row.opName}</SizableText>
-                {row.iter !== null ? (
-                  <XStack gap="$1" alignItems="center">
-                    {row.isActualDefaultIter ? (
-                      <SizableText size="$bodySmMedium">
-                        {CRYPTO_GALLERY_DEFAULT_PATH_EMOJI}
-                      </SizableText>
-                    ) : null}
-                    <SizableText
-                      size="$bodySmMedium"
-                      color={
-                        row.isActualDefaultIter
-                          ? '$textInteractive'
-                          : '$textSubdued'
-                      }
-                    >
-                      {row.iter.toLocaleString()}
-                    </SizableText>
-                  </XStack>
-                ) : null}
-              </Stack>
-              {kdfBackendColumns.map(({ backend }) => (
-                <Stack
-                  key={backend}
-                  flexBasis={kdfBackendColumnFlexBasis}
-                  alignItems="flex-end"
-                >
-                  {renderBackendCell(row, backend, row[backend])}
-                </Stack>
-              ))}
-            </XStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload({
-                    defaultProbe: lastRunDefaultProbe,
-                    selectedIter: lastRunSelectedIter,
-                    rows: kdfTimingTableRows,
-                  }),
-                );
-              }}
-            >
-              Copy KDF timing table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {actualOutputs &&
-      (Object.values(actualOutputs.sha256).some(Boolean) ||
-        Object.values(actualOutputs.sha512).some(Boolean)) ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <SizableText size="$bodySmMedium" color="$textSubdued">
-              PBKDF2 actual output @ iter ={' '}
-              {actualOutputs.iter.toLocaleString()} (base64 + hex preview,
-              cross-device compare)
-            </SizableText>
-          </XStack>
-          {(
-            [
-              ['sha256', 'noble', 'sha-256 noble'],
-              ['sha256', 'native', 'sha-256 native (aes-crypto)'],
-              ['sha256', 'fastNative', 'sha-256 fast (fast-pbkdf2)'],
-              ['sha256', 'quickNative', 'sha-256 quick (quick-crypto)'],
-              ['sha256', 'defaultBackend', 'sha-256 default backend'],
-              ['sha512', 'noble', 'sha-512 noble'],
-              ['sha512', 'native', 'sha-512 native (aes-crypto)'],
-              ['sha512', 'fastNative', 'sha-512 fast (fast-pbkdf2)'],
-              ['sha512', 'quickNative', 'sha-512 quick (quick-crypto)'],
-            ] as const
-          ).map(([bucket, backend, label]) => {
-            const value = (
-              actualOutputs[bucket] as Record<string, string | undefined>
-            )[backend];
-            if (!value) return null;
-            return (
-              <XStack
-                key={`${bucket}-${backend}`}
-                paddingVertical="$1.5"
-                paddingHorizontal="$3"
-                gap="$2"
-                alignItems="flex-start"
-              >
-                <Stack flexBasis="40%">
-                  <SizableText size="$bodySm" color="$textSubdued">
-                    {label}
-                  </SizableText>
-                </Stack>
-                <Stack flex={1}>
-                  <SizableText size="$bodySm" selectable>
-                    {value}
-                  </SizableText>
-                  <SizableText size="$bodySm" color="$textSubdued" selectable>
-                    hex {formatBase64AsShortHex(value)}
-                  </SizableText>
-                </Stack>
-              </XStack>
-            );
-          })}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(stringifyCryptoGalleryTablePayload(actualOutputs));
-              }}
-            >
-              Copy PBKDF2 output table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {lastRunSelectedIter !== null && tableRows.length > 0 ? (
-        <SizableText size="$bodySm" color="$textSubdued">
-          Last run: iter={lastRunSelectedIter.toLocaleString()} (actual default{' '}
-          {(
-            lastRunDefaultProbe?.pbkdf2Iterations ??
-            lastRunDefaultProbe?.payloadIterations ??
-            AES_GCM_V2_DEFAULT_ITER
-          ).toLocaleString()}{' '}
-          always included).
-        </SizableText>
-      ) : null}
-
-      {errorMessage ? (
-        <SizableText size="$bodyMd" color="$textCritical">
-          Error: {errorMessage}
-        </SizableText>
-      ) : null}
-    </PartContainer>
-  );
-}
-
-type INativeWebEmbedCryptoPerfRow = {
-  mode: 'cold/raw' | 'hot/cache';
-  opName: string;
-  iter: number | null;
-  defaultRuntime?: INativeWebEmbedDefaultRuntime;
-  expectedResult: string;
-  localResult: string | undefined;
-  webEmbedResult: string | undefined;
-  local: number | undefined;
-  webEmbed: number | undefined;
-  validation: string | undefined;
-  isCorrect: string | undefined;
-};
-
-type INativeWebEmbedDefaultRuntime = 'primary' | 'secondary';
-
-type INativeWebEmbedBatchPerfTraceSummaryRow = {
-  mode: INativeWebEmbedCryptoPerfRow['mode'];
-  source: IBatchGetPublicKeysPerfTraceEvent['source'];
-  name: string;
-  count: number;
-  totalMs: number;
-  avgMs: number;
-  maxMs: number;
-};
-
-type INativeWebEmbedBatchPerfTracePayload = {
-  mode: INativeWebEmbedCryptoPerfRow['mode'];
-  events: IBatchGetPublicKeysPerfTraceEvent[];
-  summary: INativeWebEmbedBatchPerfTraceSummaryRow[];
-};
-
-type INativeWebEmbedMnemonicPerfTraceRow = {
-  runtime: string;
-  stage: string;
-  durationMs: number;
-  result?: string;
-  validation: string;
-  isCorrect?: string;
-  metadata?: IMnemonicToSeedPerfTraceEvent['metadata'];
-};
-
-function roundPerfMs(value: number): number {
-  return Number(value.toFixed(2));
-}
-
-function detectMnemonicToSeedBackendFromTrace(
-  events: IMnemonicToSeedPerfTraceEvent[],
-): string | undefined {
-  for (let i = events.length - 1; i >= 0; i -= 1) {
-    const event = events[i];
-    if (event.name.startsWith('mnemonicToSeed.quickCrypto.pbkdf2')) {
-      return 'react-native-quick-crypto';
-    }
-    if (event.name === 'mnemonicToSeed.webCrypto.deriveBits') {
-      return 'webcrypto';
-    }
-    if (event.name === 'mnemonicToSeed.asmcrypto.pbkdf2Sync') {
-      return 'asmcrypto';
-    }
-    if (event.name.startsWith('mnemonicToSeed.bip39.noblePbkdf2')) {
-      return 'noble';
-    }
-  }
-  return undefined;
-}
-
-function getGalleryWebCryptoSubtle(): SubtleCrypto | undefined {
-  const subtle = globalThis.crypto?.subtle;
-  if (
-    subtle &&
-    typeof subtle.importKey === 'function' &&
-    typeof subtle.deriveBits === 'function'
-  ) {
-    return subtle;
-  }
-  return undefined;
-}
-
-function toGalleryWebCryptoArrayBuffer(buffer: Buffer): ArrayBuffer {
-  const arrayBuffer = new ArrayBuffer(buffer.length);
-  const bytes = new Uint8Array(arrayBuffer);
-  bytes.set(buffer);
-  return arrayBuffer;
-}
-
-function buildBatchPerfTraceSummary({
-  mode,
-  events,
-}: {
-  mode: INativeWebEmbedCryptoPerfRow['mode'];
-  events: IBatchGetPublicKeysPerfTraceEvent[];
-}): INativeWebEmbedBatchPerfTraceSummaryRow[] {
-  const summaryMap = new Map<
-    string,
-    Omit<INativeWebEmbedBatchPerfTraceSummaryRow, 'avgMs'>
-  >();
-  for (const event of events) {
-    const key = `${event.source}:${event.name}`;
-    const current = summaryMap.get(key);
-    if (current) {
-      current.count += 1;
-      current.totalMs += event.durationMs;
-      current.maxMs = Math.max(current.maxMs, event.durationMs);
-    } else {
-      summaryMap.set(key, {
-        mode,
-        source: event.source,
-        name: event.name,
-        count: 1,
-        totalMs: event.durationMs,
-        maxMs: event.durationMs,
-      });
-    }
-  }
-  return Array.from(summaryMap.values())
-    .map((row) => ({
-      ...row,
-      totalMs: roundPerfMs(row.totalMs),
-      avgMs: roundPerfMs(row.totalMs / row.count),
-      maxMs: roundPerfMs(row.maxMs),
-    }))
-    .toSorted((a, b) => b.totalMs - a.totalMs)
-    .slice(0, 24);
-}
-
-async function runCryptoPerfTask({
-  name,
-  expect,
-  fn,
-}: {
-  name: string;
-  expect: string;
-  fn: () => Promise<Buffer | string> | Buffer | string;
-}): Promise<IRunAppCryptoTestTaskResult> {
-  const start = Date.now();
-  let result: string | undefined = '';
-  let error: string | undefined = '';
-  let isPromise = false;
-  try {
-    const value = fn();
-    if (value instanceof Promise) {
-      result = bufferUtils.bytesToHex(await value);
-      isPromise = true;
-    } else {
-      result = bufferUtils.bytesToHex(value);
-    }
-  } catch (e) {
-    error = (e as Error | undefined)?.message ?? 'Error';
-  }
-  const time = Date.now() - start;
-  return {
-    name,
-    result: result || undefined,
-    time,
-    isSlow: time > 10 && !error ? AppCryptoTestEmoji.isSlow : undefined,
-    isPromise,
-    ERROR: error ? `${AppCryptoTestEmoji.isWarning} ${error}` : undefined,
-    isCorrect: (() => {
-      if (result === expect) return AppCryptoTestEmoji.isCorrect;
-      if (error) return AppCryptoTestEmoji.isWarning;
-      return AppCryptoTestEmoji.isIncorrect;
-    })(),
-  };
-}
-
-function NativeWebEmbedCryptoPerfTest() {
-  const [resultJson, setResultJson] = useState('');
-  const [tableRows, setTableRows] = useState<INativeWebEmbedCryptoPerfRow[]>(
-    [],
-  );
-  const [batchPerfTraceRows, setBatchPerfTraceRows] = useState<
-    INativeWebEmbedBatchPerfTraceSummaryRow[]
-  >([]);
-  const [mnemonicPerfTraceRows, setMnemonicPerfTraceRows] = useState<
-    INativeWebEmbedMnemonicPerfTraceRow[]
-  >([]);
-  const [errorMessage, setErrorMessage] = useState('');
-  const [running, setRunning] = useState(false);
-  const { copyText } = useClipboard();
-  const primaryRuntimeName = platformEnv.isNative ? 'native' : 'asmcrypto';
-  const secondaryRuntimeName = platformEnv.isNative ? 'web-embed' : 'webcrypto';
-  const defaultPathEmoji = CRYPTO_GALLERY_DEFAULT_PATH_EMOJI;
-
-  const fmtCell = (v: number | undefined) =>
-    v === undefined ? '—' : `${v} ms`;
-  const fmtResultCell = (v: string | undefined) => v ?? '-';
-  const msColor = (v: number | undefined): string => {
-    if (v === undefined) return '$textDisabled';
-    if (v <= 10) return '$textSuccess';
-    if (v < 100) return '$text';
-    if (v < 500) return '$textCaution';
-    return '$textCritical';
-  };
-  const validationColor = (isCorrect: string | undefined): string => {
-    if (isCorrect === AppCryptoTestEmoji.isCorrect) {
-      return '$textSuccess';
-    }
-    if (isCorrect === AppCryptoTestEmoji.isWarning) {
-      return '$textCaution';
-    }
-    return '$textCritical';
-  };
-  const fmtMetadata = (
-    metadata: IMnemonicToSeedPerfTraceEvent['metadata'],
-  ): string =>
-    metadata
-      ? Object.entries(metadata)
-          .map(([key, value]) => `${key}=${String(value)}`)
-          .join(', ')
-      : '';
-  const getDefaultRuntimeByBackend = (
-    backend: string | undefined,
-  ): INativeWebEmbedDefaultRuntime | undefined => {
-    if (!backend) {
-      return undefined;
-    }
-    if (backend === primaryRuntimeName) {
-      return 'primary';
-    }
-    if (backend === secondaryRuntimeName) {
-      return 'secondary';
-    }
-    if (platformEnv.isNative && backend.startsWith('react-native-')) {
-      return 'primary';
-    }
-    if (!platformEnv.isNative && backend === 'asmcrypto') {
-      return 'primary';
-    }
-    if (!platformEnv.isNative && backend === 'webcrypto') {
-      return 'secondary';
-    }
-    return undefined;
-  };
-  const renderRuntimeCell = ({
-    row,
-    runtime,
-    value,
-  }: {
-    row: INativeWebEmbedCryptoPerfRow;
-    runtime: INativeWebEmbedDefaultRuntime;
-    value: number | undefined;
-  }) => (
-    <XStack gap="$1" alignItems="center" justifyContent="flex-end">
-      {row.defaultRuntime === runtime ? (
-        <SizableText size="$bodyMd">{defaultPathEmoji}</SizableText>
-      ) : null}
-      <SizableText size="$bodyMd" color={msColor(value)}>
-        {fmtCell(value)}
-      </SizableText>
-    </XStack>
-  );
-
-  const testNativeWebEmbedPerf = async () => {
-    try {
-      const secret = await loadCoreSecret();
-      const {
-        batchGetPublicKeys,
-        clearHdCredentialDecryptCache,
-        clearPbkdf2CacheAsync,
-        encryptRevealableSeed,
-        mnemonicToRevealableSeed,
-        mnemonicToSeedAsync,
-      } = secret;
-      const tasks: IRunAppCryptoTestTaskResult[] = [];
-      const rows: INativeWebEmbedCryptoPerfRow[] = [];
-      const batchPerfTraces: INativeWebEmbedBatchPerfTracePayload[] = [];
-      const nextBatchPerfTraceRows: INativeWebEmbedBatchPerfTraceSummaryRow[] =
-        [];
-      const mnemonicPerfEvents: IMnemonicToSeedPerfTraceEvent[] = [];
-      const nextMnemonicPerfTraceRows: INativeWebEmbedMnemonicPerfTraceRow[] =
-        [];
-      const webEmbedApiStatus = globalThis as {
-        $onekeyAppWebembedApiWebviewInitFailed?: boolean;
-      };
-      const webEmbedApiWasMarkedFailed = Boolean(
-        webEmbedApiStatus.$onekeyAppWebembedApiWebviewInitFailed,
-      );
-      const canUseWebEmbed = platformEnv.isNative && !platformEnv.isJest;
-      const nonDbTxKdfParams = appCrypto.pbkdf2.getPbkdf2KdfParamsForNonDbTx();
-      const canUseWebCryptoKdf = nonDbTxKdfParams.kdfBackend === 'webcrypto';
-      let canRunSecondary = platformEnv.isNative
-        ? canUseWebEmbed
-        : canUseWebCryptoKdf;
-      const webEmbedPrewarmResults: {
-        attempt: number;
-        durationMs: number;
-        error?: string;
-        result?: string;
-        status: 'failed' | 'skipped' | 'success';
-      }[] = [];
-      let shouldCooldownBeforeBenchmarkTask = false;
-      const waitBeforeBenchmarkTask = async () => {
-        if (!shouldCooldownBeforeBenchmarkTask) {
-          shouldCooldownBeforeBenchmarkTask = true;
-          return;
-        }
-        await waitForAesGcmV2BenchmarkCooldown();
-      };
-      const runPair = async ({
-        mode,
-        opName,
-        iter,
-        expect,
-        beforeLocalFn,
-        localFn,
-        beforeSecondaryFn,
-        secondaryFn,
-        defaultRuntime,
-      }: {
-        mode: INativeWebEmbedCryptoPerfRow['mode'];
-        opName: string;
-        iter: number | null;
-        expect: string;
-        beforeLocalFn?: () => Promise<void>;
-        localFn: () => Promise<Buffer | string> | Buffer | string;
-        beforeSecondaryFn?: () => Promise<void>;
-        secondaryFn: () => Promise<Buffer | string> | Buffer | string;
-        defaultRuntime?: INativeWebEmbedDefaultRuntime;
-      }) => {
-        await waitBeforeBenchmarkTask();
-        await beforeLocalFn?.();
-        if (beforeLocalFn) {
-          await waitBeforeBenchmarkTask();
-        }
-        const localTask = await runCryptoPerfTask({
-          name: `${mode} ${opName} ${primaryRuntimeName}`,
-          expect,
-          fn: localFn,
-        });
-        tasks.push(localTask);
-
-        let secondaryTask: IRunAppCryptoTestTaskResult | undefined;
-        if (canRunSecondary) {
-          await waitBeforeBenchmarkTask();
-          await beforeSecondaryFn?.();
-          if (beforeSecondaryFn) {
-            await waitBeforeBenchmarkTask();
-          }
-          secondaryTask = await runCryptoPerfTask({
-            name: `${mode} ${opName} ${secondaryRuntimeName}`,
-            expect,
-            fn: secondaryFn,
-          });
-          if (
-            !secondaryTask.ERROR &&
-            secondaryTask.result !== localTask.result
-          ) {
-            secondaryTask.isCorrect = AppCryptoTestEmoji.isIncorrect;
-          }
-          tasks.push(secondaryTask);
-        }
-        const rowValidation = (() => {
-          if (localTask.ERROR) return localTask.ERROR;
-          if (localTask.result !== expect) {
-            return `${primaryRuntimeName} result mismatch`;
-          }
-          if (!secondaryTask) return `${primaryRuntimeName} result verified`;
-          if (secondaryTask.ERROR) return secondaryTask.ERROR;
-          if (secondaryTask.result !== expect) {
-            return `${secondaryRuntimeName} result mismatch`;
-          }
-          if (secondaryTask.result !== localTask.result) {
-            return `${primaryRuntimeName}/${secondaryRuntimeName} result mismatch`;
-          }
-          return `${primaryRuntimeName}/${secondaryRuntimeName} matched`;
-        })();
-        const rowIsCorrect = (() => {
-          if (!secondaryTask) {
-            return AppCryptoTestEmoji.isWarning;
-          }
-          return localTask.isCorrect === AppCryptoTestEmoji.isCorrect &&
-            secondaryTask.isCorrect === AppCryptoTestEmoji.isCorrect
-            ? AppCryptoTestEmoji.isCorrect
-            : AppCryptoTestEmoji.isIncorrect;
-        })();
-        rows.push({
-          mode,
-          opName,
-          iter,
-          defaultRuntime,
-          expectedResult: expect,
-          localResult: localTask.result,
-          webEmbedResult: secondaryTask?.result,
-          local: localTask.time,
-          webEmbed: secondaryTask?.time,
-          validation: rowValidation,
-          isCorrect: rowIsCorrect,
-        });
-        return {
-          localTask,
-          secondaryTask,
-        };
-      };
-
-      const password = 'onekey-gallery-password';
-      const mnemonicPerfTrace: IMnemonicToSeedPerfTrace = {
-        onEvent: (event) => {
-          const tracedEvent: IMnemonicToSeedPerfTraceEvent = {
-            ...event,
-            durationMs: roundPerfMs(event.durationMs),
-          };
-          mnemonicPerfEvents.push(tracedEvent);
-          nextMnemonicPerfTraceRows.push({
-            runtime: primaryRuntimeName,
-            stage: tracedEvent.name,
-            durationMs: tracedEvent.durationMs,
-            metadata: tracedEvent.metadata,
-            validation: 'timed',
-          });
-        },
-      };
-      const runLocalBatchGetPublicKeysWithTrace = async ({
-        mode,
-        fn,
-      }: {
-        mode: INativeWebEmbedCryptoPerfRow['mode'];
-        fn: (perfTrace: IBatchGetPublicKeysPerfTrace) => Promise<Buffer>;
-      }): Promise<Buffer> => {
-        const events: IBatchGetPublicKeysPerfTraceEvent[] = [];
-        const perfTrace: IBatchGetPublicKeysPerfTrace = {
-          onEvent: (event) => {
-            events.push({
-              ...event,
-              durationMs: roundPerfMs(event.durationMs),
-            });
-          },
-        };
-        const result = await fn(perfTrace);
-        const summary = buildBatchPerfTraceSummary({
-          mode,
-          events,
-        });
-        batchPerfTraces.push({
-          mode,
-          events,
-          summary,
-        });
-        nextBatchPerfTraceRows.push(...summary);
-        return result;
-      };
-      const prewarmWebEmbedApi = async () => {
-        if (!platformEnv.isNative || !canUseWebEmbed) {
-          webEmbedPrewarmResults.push({
-            attempt: 0,
-            durationMs: 0,
-            status: 'skipped',
-          });
-          return true;
-        }
-        const maxAttempts = 2;
-        for (let attempt = 1; attempt <= maxAttempts; attempt += 1) {
-          await waitBeforeBenchmarkTask();
-          const start = Date.now();
-          try {
-            const webembedApiProxy = (
-              await import('@onekeyhq/kit-bg/src/webembeds/instance/webembedApiProxy')
-            ).default;
-            const result = await webembedApiProxy.test.test1(
-              'crypto-gallery-prewarm',
-              String(attempt),
-            );
-            const durationMs = Date.now() - start;
-            webEmbedApiStatus.$onekeyAppWebembedApiWebviewInitFailed = false;
-            webEmbedPrewarmResults.push({
-              attempt,
-              durationMs,
-              result,
-              status: 'success',
-            });
-            nextMnemonicPerfTraceRows.push({
-              runtime: secondaryRuntimeName,
-              stage: 'webEmbed.prewarm',
-              durationMs,
-              result,
-              metadata: {
-                attempt,
-                wasMarkedFailed: webEmbedApiWasMarkedFailed,
-              },
-              validation: 'ready',
-              isCorrect: AppCryptoTestEmoji.isCorrect,
-            });
-            return true;
-          } catch (error) {
-            const durationMs = Date.now() - start;
-            const prewarmErrorMessage = (error as Error).message;
-            webEmbedPrewarmResults.push({
-              attempt,
-              durationMs,
-              error: prewarmErrorMessage,
-              status: 'failed',
-            });
-            nextMnemonicPerfTraceRows.push({
-              runtime: secondaryRuntimeName,
-              stage: 'webEmbed.prewarm',
-              durationMs,
-              metadata: {
-                attempt,
-                wasMarkedFailed: webEmbedApiWasMarkedFailed,
-              },
-              validation: prewarmErrorMessage,
-              isCorrect: AppCryptoTestEmoji.isWarning,
-            });
-          }
-        }
-        return false;
-      };
-      const clearColdCaches = async () => {
-        await clearPbkdf2CacheAsync();
-      };
-      const testMnemonic =
-        'test test test test test test test test test test test junk';
-      const testPassphrase = 'optional passphrase';
-      const mnemonicToSeedExpected =
-        'bc0d03ab4f8871dd4a7a68423894bb88fb54973899e4721c9dffd09a5b589171b5712b27da764f7be0653ba361f445b4f9251b490525833b644b7a13eebc7e2c';
-      const batchGetPublicKeysExpected =
-        '02be3c6a2aa12821e037632d30ed51a6dd2a41f2277959d56db41dfb94439e49e9';
-      const defaultPathProbe: Record<
-        string,
-        {
-          backend?: string;
-          runtime?: INativeWebEmbedDefaultRuntime;
-        }
-      > = {};
-      const bip39Pbkdf2Metadata = {
-        iterations: 2048,
-        keyLength: 64,
-        digest: 'sha512',
-      };
-      const probeDefaultMnemonicToSeedBackend = async () => {
-        await waitBeforeBenchmarkTask();
-        const probeEvents: IMnemonicToSeedPerfTraceEvent[] = [];
-        const result = await mnemonicToSeedAsync({
-          mnemonic: testMnemonic,
-          passphrase: testPassphrase,
-          useWebembedApi: false,
-          perfTrace: {
-            onEvent: (event) => {
-              probeEvents.push(event);
-            },
-          },
-        });
-        const resultHex = bufferUtils.bytesToHex(result);
-        if (resultHex !== mnemonicToSeedExpected) {
-          throw new OneKeyLocalError(
-            'mnemonicToSeedAsync default probe result mismatch',
-          );
-        }
-        return detectMnemonicToSeedBackendFromTrace(probeEvents);
-      };
-      const runMnemonicSegmentDiagnostic = async ({
-        runtime,
-        stage,
-        fn,
-      }: {
-        runtime: string;
-        stage: string;
-        fn: () => Promise<Buffer> | Buffer;
-      }) => {
-        await waitBeforeBenchmarkTask();
-        const task = await runCryptoPerfTask({
-          name: `mnemonicToSeed segment ${stage} ${runtime}`,
-          expect: mnemonicToSeedExpected,
-          fn,
-        });
-        tasks.push(task);
-        nextMnemonicPerfTraceRows.push({
-          runtime,
-          stage,
-          durationMs: task.time,
-          result: task.result,
-          metadata: bip39Pbkdf2Metadata,
-          validation:
-            task.isCorrect === AppCryptoTestEmoji.isCorrect
-              ? 'matches expected seed'
-              : task.ERROR || 'result mismatch',
-          isCorrect: task.isCorrect,
-        });
-      };
-      const defaultMnemonicBackend = await probeDefaultMnemonicToSeedBackend();
-      const defaultMnemonicRuntime = getDefaultRuntimeByBackend(
-        defaultMnemonicBackend,
-      );
-      defaultPathProbe.mnemonicToSeedAsync = {
-        backend: defaultMnemonicBackend,
-        runtime: defaultMnemonicRuntime,
-      };
-      if (platformEnv.isNative && canRunSecondary) {
-        canRunSecondary = await prewarmWebEmbedApi();
-      }
-      const mnemonicPair = await runPair({
-        mode: 'cold/raw',
-        opName: 'mnemonicToSeedAsync',
-        iter: 2048,
-        expect: mnemonicToSeedExpected,
-        defaultRuntime: defaultMnemonicRuntime,
-        localFn: () =>
-          mnemonicToSeedAsync({
-            mnemonic: testMnemonic,
-            passphrase: testPassphrase,
-            perfTrace: mnemonicPerfTrace,
-            kdfBackend: platformEnv.isNative ? undefined : 'asmcrypto',
-            useWebembedApi: false,
-          }),
-        secondaryFn: () =>
-          mnemonicToSeedAsync({
-            mnemonic: testMnemonic,
-            passphrase: testPassphrase,
-            kdfBackend: platformEnv.isNative ? undefined : 'webcrypto',
-            useWebembedApi: platformEnv.isNative,
-          }),
-      });
-      if (mnemonicPair.secondaryTask) {
-        nextMnemonicPerfTraceRows.push({
-          runtime: secondaryRuntimeName,
-          stage: 'mnemonicToSeedAsync.total',
-          durationMs: mnemonicPair.secondaryTask.time,
-          result: mnemonicPair.secondaryTask.result,
-          metadata: bip39Pbkdf2Metadata,
-          validation:
-            mnemonicPair.secondaryTask.isCorrect ===
-            AppCryptoTestEmoji.isCorrect
-              ? 'matches expected seed'
-              : mnemonicPair.secondaryTask.ERROR || 'result mismatch',
-          isCorrect: mnemonicPair.secondaryTask.isCorrect,
-        });
-      } else if (platformEnv.isNative) {
-        nextMnemonicPerfTraceRows.push({
-          runtime: secondaryRuntimeName,
-          stage: 'mnemonicToSeedAsync.total',
-          durationMs: 0,
-          metadata: {
-            reason: 'web-embed unavailable',
-          },
-          validation: 'web-embed unavailable',
-          isCorrect: AppCryptoTestEmoji.isWarning,
-        });
-      }
-
-      const mnemonicBuffer = Buffer.from(
-        testMnemonic.normalize('NFKD'),
-        'utf8',
-      );
-      const saltBuffer = Buffer.from(
-        `mnemonic${testPassphrase.normalize('NFKD')}`,
-        'utf8',
-      );
-      const galleryWebCryptoSubtle = getGalleryWebCryptoSubtle();
-      if (galleryWebCryptoSubtle) {
-        await runMnemonicSegmentDiagnostic({
-          runtime: 'web-crypto',
-          stage: 'rawPbkdf2DeriveBits',
-          fn: async () => {
-            const key = await galleryWebCryptoSubtle.importKey(
-              'raw',
-              toGalleryWebCryptoArrayBuffer(mnemonicBuffer),
-              'PBKDF2',
-              false,
-              ['deriveBits'],
-            );
-            const derivedBits = await galleryWebCryptoSubtle.deriveBits(
-              {
-                name: 'PBKDF2',
-                salt: toGalleryWebCryptoArrayBuffer(saltBuffer),
-                iterations: 2048,
-                hash: 'SHA-512',
-              },
-              key,
-              64 * 8,
-            );
-            return Buffer.from(derivedBits);
-          },
-        });
-      }
-      await runMnemonicSegmentDiagnostic({
-        runtime: 'noble-js',
-        stage: 'rawPbkdf2Async',
-        fn: async () =>
-          Buffer.from(
-            await nobleP2Async(nobleSha512, mnemonicBuffer, saltBuffer, {
-              c: 2048,
-              dkLen: 64,
-            }),
-          ),
-      });
-      if (platformEnv.isNative) {
-        await runMnemonicSegmentDiagnostic({
-          runtime: 'quick-crypto',
-          stage: 'rawPbkdf2Async',
-          fn: () =>
-            new Promise<Buffer>((resolve, reject) => {
-              RN_QUICK_CRYPTO.pbkdf2(
-                mnemonicBuffer,
-                saltBuffer,
-                2048,
-                64,
-                'sha512',
-                (err, seed) => {
-                  if (err || !seed) {
-                    reject(
-                      err ||
-                        new OneKeyLocalError(
-                          'quick-crypto mnemonic segment failed',
-                        ),
-                    );
-                    return;
-                  }
-                  resolve(Buffer.from(seed));
-                },
-              );
-            }),
-        });
-        await runMnemonicSegmentDiagnostic({
-          runtime: 'quick-crypto',
-          stage: 'rawPbkdf2Sync',
-          fn: () =>
-            Buffer.from(
-              RN_QUICK_CRYPTO.pbkdf2Sync(
-                mnemonicBuffer,
-                saltBuffer,
-                2048,
-                64,
-                'sha512',
-              ),
-            ),
-        });
-      }
-
-      const rs = mnemonicToRevealableSeed(testMnemonic, testPassphrase);
-      const encodedPassword =
-        await backgroundApiProxy.servicePassword.encodeSensitiveText({
-          text: password,
-        });
-      const hdCredential = await encryptRevealableSeed({
-        rs,
-        password: encodedPassword,
-      });
-      const localHotScopeId = `gallery-${primaryRuntimeName}-hot:${Date.now()}`;
-      const secondaryHotScopeId = `gallery-${secondaryRuntimeName}-hot:${Date.now()}`;
-      const curveName: ICurveName = 'secp256k1';
-      const prefix = 'm';
-      const relPaths = ['0/0', '0/1', "44'/0'/0'/0/0"];
-      const primaryBatchGetPublicKeysKdfParams = platformEnv.isNative
-        ? nonDbTxKdfParams
-        : {};
-      const secondaryBatchGetPublicKeysKdfParams = platformEnv.isNative
-        ? {}
-        : nonDbTxKdfParams;
-      const batchGetPublicKeysLocalRaw = async (
-        perfTrace?: IBatchGetPublicKeysPerfTrace,
-      ) => {
-        const r = await batchGetPublicKeys({
-          curveName,
-          hdCredential,
-          password: encodedPassword,
-          prefix,
-          relPaths,
-          perfTrace,
-          useWebembedApi: false,
-          ...primaryBatchGetPublicKeysKdfParams,
-        });
-        const key = r?.[2]?.extendedKey?.key;
-        if (!key) {
-          throw new OneKeyLocalError('batchGetPublicKeys result missing');
-        }
-        return key;
-      };
-      const batchGetPublicKeysWebEmbedRaw = async () => {
-        const r = await batchGetPublicKeys({
-          curveName,
-          hdCredential,
-          password: encodedPassword,
-          prefix,
-          relPaths,
-          useWebembedApi: platformEnv.isNative,
-          ...secondaryBatchGetPublicKeysKdfParams,
-        });
-        const key = r?.[2]?.extendedKey?.key;
-        if (!key) {
-          throw new OneKeyLocalError('batchGetPublicKeys result missing');
-        }
-        return key;
-      };
-      const batchGetPublicKeysLocalHot = async (
-        perfTrace?: IBatchGetPublicKeysPerfTrace,
-      ) => {
-        const r = await batchGetPublicKeys({
-          curveName,
-          hdCredential,
-          password: encodedPassword,
-          prefix,
-          relPaths,
-          hdCredentialCacheScopeId: localHotScopeId,
-          perfTrace,
-          useWebembedApi: false,
-          ...primaryBatchGetPublicKeysKdfParams,
-        });
-        const key = r?.[2]?.extendedKey?.key;
-        if (!key) {
-          throw new OneKeyLocalError('batchGetPublicKeys result missing');
-        }
-        return key;
-      };
-      const batchGetPublicKeysWebEmbedHot = async () => {
-        const r = await batchGetPublicKeys({
-          curveName,
-          hdCredential,
-          password: encodedPassword,
-          prefix,
-          relPaths,
-          hdCredentialCacheScopeId: secondaryHotScopeId,
-          useWebembedApi: platformEnv.isNative,
-          ...secondaryBatchGetPublicKeysKdfParams,
-        });
-        const key = r?.[2]?.extendedKey?.key;
-        if (!key) {
-          throw new OneKeyLocalError('batchGetPublicKeys result missing');
-        }
-        return key;
-      };
-      const probeDefaultBatchGetPublicKeysBackend = async () => {
-        await waitBeforeBenchmarkTask();
-        await clearColdCaches();
-        const debugCryptoProbeId = `crypto-gallery-batch-default-${Date.now()}`;
-        appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(debugCryptoProbeId);
-        try {
-          const r = await batchGetPublicKeys({
-            curveName,
-            hdCredential,
-            password: encodedPassword,
-            prefix,
-            relPaths,
-            useWebembedApi: false,
-            debugCryptoProbeId,
-            ...nonDbTxKdfParams,
-          });
-          const key = r?.[2]?.extendedKey?.key;
-          if (
-            !key ||
-            bufferUtils.bytesToHex(key) !== batchGetPublicKeysExpected
-          ) {
-            throw new OneKeyLocalError(
-              'batchGetPublicKeys default probe result mismatch',
-            );
-          }
-          return appCrypto.pbkdf2.getPbkdf2InvocationByProbeId(
-            debugCryptoProbeId,
-          )?.backend;
-        } finally {
-          appCrypto.pbkdf2.clearPbkdf2InvocationByProbeId(debugCryptoProbeId);
-        }
-      };
-      const defaultBatchGetPublicKeysBackend =
-        await probeDefaultBatchGetPublicKeysBackend();
-      const defaultBatchGetPublicKeysRuntime = getDefaultRuntimeByBackend(
-        defaultBatchGetPublicKeysBackend,
-      );
-      defaultPathProbe.batchGetPublicKeys = {
-        backend: defaultBatchGetPublicKeysBackend,
-        runtime: defaultBatchGetPublicKeysRuntime,
-      };
-      try {
-        await runPair({
-          mode: 'cold/raw',
-          opName: 'batchGetPublicKeys',
-          iter: null,
-          expect: batchGetPublicKeysExpected,
-          defaultRuntime: defaultBatchGetPublicKeysRuntime,
-          beforeLocalFn: clearColdCaches,
-          localFn: () =>
-            runLocalBatchGetPublicKeysWithTrace({
-              mode: 'cold/raw',
-              fn: batchGetPublicKeysLocalRaw,
-            }),
-          beforeSecondaryFn: clearColdCaches,
-          secondaryFn: batchGetPublicKeysWebEmbedRaw,
-        });
-        await runPair({
-          mode: 'hot/cache',
-          opName: 'batchGetPublicKeys',
-          iter: null,
-          expect: batchGetPublicKeysExpected,
-          defaultRuntime: defaultBatchGetPublicKeysRuntime,
-          beforeLocalFn: async () => {
-            await clearHdCredentialDecryptCache({
-              hdCredentialCacheScopeId: localHotScopeId,
-            });
-            await batchGetPublicKeysLocalHot();
-          },
-          localFn: () =>
-            runLocalBatchGetPublicKeysWithTrace({
-              mode: 'hot/cache',
-              fn: batchGetPublicKeysLocalHot,
-            }),
-          beforeSecondaryFn: async () => {
-            await clearHdCredentialDecryptCache({
-              hdCredentialCacheScopeId: secondaryHotScopeId,
-            });
-            await batchGetPublicKeysWebEmbedHot();
-          },
-          secondaryFn: batchGetPublicKeysWebEmbedHot,
-        });
-      } finally {
-        await clearHdCredentialDecryptCache({
-          hdCredentialCacheScopeId: localHotScopeId,
-        });
-        await clearHdCredentialDecryptCache({
-          hdCredentialCacheScopeId: secondaryHotScopeId,
-        });
-      }
-
-      setTableRows(rows);
-      setBatchPerfTraceRows(nextBatchPerfTraceRows);
-      setMnemonicPerfTraceRows(nextMnemonicPerfTraceRows);
-      setResultJson(
-        stringifyCryptoGalleryTablePayload({
-          platform: {
-            isNative: platformEnv.isNative,
-            isNativeIOS: platformEnv.isNativeIOS,
-            isNativeAndroid: platformEnv.isNativeAndroid,
-            canUseWebEmbed,
-            webEmbedApiWasMarkedFailed,
-            primaryRuntime: primaryRuntimeName,
-            secondaryRuntime: secondaryRuntimeName,
-            canUseWebCryptoKdf,
-          },
-          defaultPath: {
-            pbkdf2: appCrypto.pbkdf2.getPbkdf2BackendForCurrentPlatform(),
-            aesGcm: appCrypto.aesGcm.getAesGcmBackendForCurrentPlatform(),
-            actualProbe: defaultPathProbe,
-            primaryAccountDerivationKdfParams:
-              primaryBatchGetPublicKeysKdfParams,
-            secondaryAccountDerivationKdfParams:
-              secondaryBatchGetPublicKeysKdfParams,
-          },
-          expectedResults: {
-            mnemonicToSeedAsync: mnemonicToSeedExpected,
-            batchGetPublicKeys: batchGetPublicKeysExpected,
-          },
-          resultRows: rows,
-          mnemonicToSeedLocalPerfTrace: mnemonicPerfEvents,
-          mnemonicToSeedSegmentedValidation: nextMnemonicPerfTraceRows,
-          batchGetPublicKeysLocalPerfTrace: batchPerfTraces,
-          webEmbedPrewarmResults,
-          tasks,
-        }),
-      );
-      setErrorMessage('');
-
-      const allPassed =
-        tasks.every((t) => t.isCorrect === AppCryptoTestEmoji.isCorrect) &&
-        rows.every((row) => row.isCorrect === AppCryptoTestEmoji.isCorrect);
-      if (allPassed) {
-        setErrorMessage('');
-        Toast.success({
-          title: 'Secret Functions Crypto Perf (v2 test) passed',
-        });
-      } else {
-        const failedTaskNames = tasks
-          .filter((t) => t.isCorrect !== AppCryptoTestEmoji.isCorrect)
-          .map((t) => t.name)
-          .concat(
-            rows
-              .filter((row) => row.isCorrect !== AppCryptoTestEmoji.isCorrect)
-              .map((row) => `${row.mode} ${row.opName}`),
-          )
-          .slice(0, 4);
-        setErrorMessage(
-          `Validation failed: ${failedTaskNames.join(', ')}${
-            failedTaskNames.length >= 4 ? '...' : ''
-          }`,
-        );
-        Toast.error({
-          title: 'Secret Functions Crypto Perf (v2 test) failed',
-        });
-      }
-    } catch (error) {
-      setErrorMessage((error as Error).message);
-      setResultJson('');
-      setTableRows([]);
-      setBatchPerfTraceRows([]);
-      setMnemonicPerfTraceRows([]);
-      Toast.error({
-        title: `Secret Functions Crypto Perf (v2 test) failed: ${
-          (error as Error).message
-        }`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="Secret Functions Crypto Perf (v2 test)">
-      <XStack gap="$3" alignItems="center" flexWrap="wrap">
-        <Button
-          variant="primary"
-          loading={running}
-          disabled={running}
-          onPress={async () => {
-            setRunning(true);
-            try {
-              await runCryptoGalleryTestExclusive(testNativeWebEmbedPerf);
-            } finally {
-              setRunning(false);
-            }
-          }}
-        >
-          Run Test
-        </Button>
-        {resultJson ? (
-          <Button
-            size="small"
-            onPress={() => {
-              copyText(resultJson);
-            }}
-          >
-            Copy raw JSON
-          </Button>
-        ) : null}
-      </XStack>
-      <SizableText size="$bodySm" color="$textSubdued">
-        Compares mnemonicToSeedAsync and batchGetPublicKeys on the current local
-        runtime.
-      </SizableText>
-
-      {tableRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <Stack flexBasis="34%">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                operation
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="22%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                {primaryRuntimeName}
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="22%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                {secondaryRuntimeName}
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="22%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                validation
-              </SizableText>
-            </Stack>
-          </XStack>
-          {tableRows.map((row, idx) => (
-            <XStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={idx === tableRows.length - 1 ? 0 : 1}
-              borderBottomColor="$borderSubdued"
-              alignItems="center"
-            >
-              <Stack flexBasis="34%">
-                <SizableText size="$bodyMd">{row.opName}</SizableText>
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {row.mode}
-                  {row.iter !== null ? ` - ${row.iter.toLocaleString()}` : ''}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="22%" alignItems="flex-end">
-                {renderRuntimeCell({
-                  row,
-                  runtime: 'primary',
-                  value: row.local,
-                })}
-              </Stack>
-              <Stack flexBasis="22%" alignItems="flex-end">
-                {renderRuntimeCell({
-                  row,
-                  runtime: 'secondary',
-                  value: row.webEmbed,
-                })}
-              </Stack>
-              <Stack flexBasis="22%" alignItems="flex-end">
-                <SizableText
-                  size="$bodySmMedium"
-                  color={validationColor(row.isCorrect)}
-                  textAlign="right"
-                >
-                  {row.isCorrect}
-                </SizableText>
-              </Stack>
-            </XStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload({
-                    primaryRuntime: primaryRuntimeName,
-                    secondaryRuntime: secondaryRuntimeName,
-                    rows: tableRows.map(
-                      ({
-                        defaultRuntime,
-                        isCorrect,
-                        iter,
-                        local,
-                        mode,
-                        opName,
-                        validation,
-                        webEmbed,
-                      }) => ({
-                        defaultRuntime,
-                        isCorrect,
-                        iter,
-                        local,
-                        mode,
-                        opName,
-                        validation,
-                        webEmbed,
-                      }),
-                    ),
-                  }),
-                );
-              }}
-            >
-              Copy overview table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {tableRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <SizableText size="$bodySmMedium" color="$textSubdued">
-              calculation results (full hex, cross-platform compare)
-            </SizableText>
-          </XStack>
-          {tableRows.map((row, idx) => (
-            <YStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              gap="$2"
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={idx === tableRows.length - 1 ? 0 : 1}
-              borderBottomColor="$borderSubdued"
-            >
-              <XStack gap="$2" alignItems="center" flexWrap="wrap">
-                <SizableText size="$bodyMd">{row.opName}</SizableText>
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {row.mode}
-                  {row.iter !== null ? ` - ${row.iter.toLocaleString()}` : ''}
-                </SizableText>
-              </XStack>
-              <YStack gap="$1">
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  expected
-                </SizableText>
-                <SizableText
-                  size="$bodySm"
-                  fontFamily="$monoRegular"
-                  selectable
-                  style={{ wordBreak: 'break-all' }}
-                >
-                  {row.expectedResult}
-                </SizableText>
-              </YStack>
-              <YStack gap="$1">
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {primaryRuntimeName} result
-                </SizableText>
-                <SizableText
-                  size="$bodySm"
-                  fontFamily="$monoRegular"
-                  selectable
-                  style={{ wordBreak: 'break-all' }}
-                >
-                  {fmtResultCell(row.localResult)}
-                </SizableText>
-              </YStack>
-              <YStack gap="$1">
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {secondaryRuntimeName} result
-                </SizableText>
-                <SizableText
-                  size="$bodySm"
-                  color={row.webEmbedResult ? '$text' : '$textDisabled'}
-                  fontFamily="$monoRegular"
-                  selectable
-                  style={{ wordBreak: 'break-all' }}
-                >
-                  {fmtResultCell(row.webEmbedResult)}
-                </SizableText>
-              </YStack>
-            </YStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload({
-                    primaryRuntime: primaryRuntimeName,
-                    secondaryRuntime: secondaryRuntimeName,
-                    rows: tableRows.map(
-                      ({
-                        expectedResult,
-                        iter,
-                        localResult,
-                        mode,
-                        opName,
-                        webEmbedResult,
-                      }) => ({
-                        expectedResult,
-                        iter,
-                        localResult,
-                        mode,
-                        opName,
-                        webEmbedResult,
-                      }),
-                    ),
-                  }),
-                );
-              }}
-            >
-              Copy calculation results table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {mnemonicPerfTraceRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <Stack flexBasis="20%">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                runtime
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="40%">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                mnemonic stage
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="18%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                time
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="22%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                validation
-              </SizableText>
-            </Stack>
-          </XStack>
-          {mnemonicPerfTraceRows.map((row, idx) => (
-            <XStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={
-                idx === mnemonicPerfTraceRows.length - 1 ? 0 : 1
-              }
-              borderBottomColor="$borderSubdued"
-              alignItems="center"
-            >
-              <Stack flexBasis="20%">
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {row.runtime}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="40%">
-                <SizableText size="$bodySm" color="$text">
-                  {row.stage}
-                </SizableText>
-                {row.metadata ? (
-                  <SizableText size="$bodySm" color="$textSubdued">
-                    {fmtMetadata(row.metadata)}
-                  </SizableText>
-                ) : null}
-              </Stack>
-              <Stack flexBasis="18%" alignItems="flex-end">
-                <SizableText size="$bodySm" color={msColor(row.durationMs)}>
-                  {fmtCell(row.durationMs)}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="22%" alignItems="flex-end">
-                <SizableText
-                  size="$bodySmMedium"
-                  color={
-                    row.isCorrect ? validationColor(row.isCorrect) : '$text'
-                  }
-                  textAlign="right"
-                >
-                  {row.isCorrect ?? ''}
-                </SizableText>
-              </Stack>
-            </XStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload(mnemonicPerfTraceRows),
-                );
-              }}
-            >
-              Copy mnemonic stage table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {batchPerfTraceRows.length > 0 ? (
-        <YStack
-          borderWidth={1}
-          borderColor="$border"
-          borderRadius="$2"
-          overflow="hidden"
-        >
-          <XStack
-            backgroundColor="$bgSubdued"
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderBottomWidth={1}
-            borderBottomColor="$border"
-          >
-            <Stack flexBasis="20%">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                mode
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="36%">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                stage
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="12%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                count
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="16%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                total
-              </SizableText>
-            </Stack>
-            <Stack flexBasis="16%" alignItems="flex-end">
-              <SizableText size="$bodySmMedium" color="$textSubdued">
-                avg
-              </SizableText>
-            </Stack>
-          </XStack>
-          {batchPerfTraceRows.map((row, idx) => (
-            <XStack
-              // eslint-disable-next-line react/no-array-index-key
-              key={idx}
-              paddingVertical="$2.5"
-              paddingHorizontal="$3"
-              borderBottomWidth={idx === batchPerfTraceRows.length - 1 ? 0 : 1}
-              borderBottomColor="$borderSubdued"
-              alignItems="center"
-            >
-              <Stack flexBasis="20%">
-                <SizableText size="$bodySmMedium" color="$textSubdued">
-                  {row.mode}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="36%">
-                <SizableText size="$bodySm" color="$text">
-                  {row.source}: {row.name}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="12%" alignItems="flex-end">
-                <SizableText size="$bodySm">{row.count}</SizableText>
-              </Stack>
-              <Stack flexBasis="16%" alignItems="flex-end">
-                <SizableText size="$bodySm" color={msColor(row.totalMs)}>
-                  {fmtCell(row.totalMs)}
-                </SizableText>
-              </Stack>
-              <Stack flexBasis="16%" alignItems="flex-end">
-                <SizableText size="$bodySm" color={msColor(row.avgMs)}>
-                  {fmtCell(row.avgMs)}
-                </SizableText>
-              </Stack>
-            </XStack>
-          ))}
-          <XStack
-            paddingVertical="$2"
-            paddingHorizontal="$3"
-            borderTopWidth={1}
-            borderTopColor="$borderSubdued"
-          >
-            <Button
-              size="small"
-              onPress={() => {
-                copyText(
-                  stringifyCryptoGalleryTablePayload(batchPerfTraceRows),
-                );
-              }}
-            >
-              Copy batch stage table
-            </Button>
-          </XStack>
-        </YStack>
-      ) : null}
-
-      {errorMessage ? (
-        <SizableText size="$bodyMd" color="$textCritical">
-          Error: {errorMessage}
-        </SizableText>
-      ) : null}
-    </PartContainer>
-  );
-}
-
-/**
- * Test crypto.subtle polyfill
- * This polyfill is required for Supabase Auth PKCE flow on React Native
- * @see packages/shared/src/appCrypto/cryptoSubtlePolyfill.js
- */
-function CryptoSubtlePolyfillTest() {
-  const [result, setResult] = useState('');
-
-  const testCryptoSubtle = async () => {
-    try {
-      const tasks: IRunAppCryptoTestTaskResult[] = [];
-
-      // Test 1: Check if crypto.subtle exists
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'true',
-          name: 'crypto.subtle exists',
-          fn: async () => {
-            return String(
-              typeof crypto !== 'undefined' &&
-                typeof crypto.subtle !== 'undefined' &&
-                typeof crypto.subtle.digest === 'function',
-            );
-          },
-        }),
-      );
-
-      // Test 2: SHA-256 digest test
-      // Hash of "hello" in SHA-256 should be:
-      // 2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824
-      const testData = new TextEncoder().encode('hello');
-      const expectedSha256 =
-        '2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824';
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: expectedSha256,
-          name: 'crypto.subtle.digest(SHA-256)',
-          fn: async () => {
-            const hashBuffer = await crypto.subtle.digest('SHA-256', testData);
-            const hashArray = Array.from(new Uint8Array(hashBuffer));
-            const hashHex = hashArray
-              .map((b) => b.toString(16).padStart(2, '0'))
-              .join('');
-            return hashHex;
-          },
-        }),
-      );
-
-      // Test 3: SHA-512 digest test
-      // Hash of "hello" in SHA-512
-      const expectedSha512 =
-        '9b71d224bd62f3785d96d46ad3ea3d73319bfbc2890caadae2dff72519673ca72323c3d99ba5c11d7c7acc6e14b8c5da0c4663475c2e5c3adef46f73bcdec043';
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: expectedSha512,
-          name: 'crypto.subtle.digest(SHA-512)',
-          fn: async () => {
-            const hashBuffer = await crypto.subtle.digest('SHA-512', testData);
-            const hashArray = Array.from(new Uint8Array(hashBuffer));
-            const hashHex = hashArray
-              .map((b) => b.toString(16).padStart(2, '0'))
-              .join('');
-            return hashHex;
-          },
-        }),
-      );
-
-      // Test 4: SHA-1 digest test
-      // Hash of "hello" in SHA-1
-      const expectedSha1 = 'aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d';
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: expectedSha1,
-          name: 'crypto.subtle.digest(SHA-1)',
-          fn: async () => {
-            const hashBuffer = await crypto.subtle.digest('SHA-1', testData);
-            const hashArray = Array.from(new Uint8Array(hashBuffer));
-            const hashHex = hashArray
-              .map((b) => b.toString(16).padStart(2, '0'))
-              .join('');
-            return hashHex;
-          },
-        }),
-      );
-
-      // Test 5: Supabase PKCE simulation test
-      // Simulate the actual PKCE flow that Supabase uses
-      const codeVerifier = 'test-code-verifier-for-pkce-flow';
-      const verifierData = new TextEncoder().encode(codeVerifier);
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            'f0c2f8b2aad90ad913c0561953b38bf3d435f59b5e4ef24eebc6605b0b444907',
-          name: 'crypto.subtle.digest(PKCE simulation)',
-          fn: async () => {
-            const hashBuffer = await crypto.subtle.digest(
-              'SHA-256',
-              verifierData,
-            );
-            const hashArray = Array.from(new Uint8Array(hashBuffer));
-            const hashHex = hashArray
-              .map((b) => b.toString(16).padStart(2, '0'))
-              .join('');
-            return hashHex;
-          },
-        }),
-      );
-
-      setResult(
-        stringUtils.stableStringify(
-          tasks,
-          stringUtils.STRINGIFY_REPLACER.bufferToHex,
-          2,
-        ),
-      );
-
-      const allPassed = tasks.every(
-        (t) => t.isCorrect === AppCryptoTestEmoji.isCorrect,
-      );
-      if (allPassed) {
-        Toast.success({
-          title: 'crypto.subtle polyfill test passed',
-        });
-      } else {
-        Toast.error({
-          title: 'crypto.subtle polyfill test failed',
-        });
-      }
-    } catch (error) {
-      setResult(`Error: ${(error as Error).message}`);
-      Toast.error({
-        title: `crypto.subtle test failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="crypto.subtle Polyfill Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testCryptoSubtle)}
-      >
-        Test crypto.subtle Polyfill
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-function SecretFunctionsTest() {
-  const [result, setResult] = useState('');
-
-  const testSecretFunctions = async () => {
-    try {
-      const {
-        batchGetPublicKeys,
-        decodeSensitiveTextAsync,
-        decryptAsync,
-        encodeSensitiveTextAsync,
-        encryptAsync,
-        encryptRevealableSeed,
-        generateRootFingerprintHexAsync,
-        mnemonicFromEntropyAsync,
-        mnemonicToRevealableSeed,
-        mnemonicToSeedAsync,
-      } = await loadCoreSecret();
-      const tasks: IRunAppCryptoTestTaskResult[] = [];
-
-      const testPasswordRaw = 'password123';
-      const encodeSensitiveTextKey =
-        'ENCODE_KEY::755174C1-6480-401A-8C3D-84ADB2E0C376::cf6e2e1c-e53b-431e-a6e4-3f27c9a7ac0b';
-
-      let customSalt = bufferUtils.toBuffer(
-        '8ff67563c060ca12aac18757221cea72482d139ea65d5f5d4f55a05c69ae87eb',
-      );
-      let customIv = bufferUtils.toBuffer('ad76f31087e49bbc59ac0f08d679e4c0');
-      // const customSalt = crypto.randomBytes(PBKDF2_SALT_LENGTH);
-      // const customIv = crypto.randomBytes(AES256_IV_LENGTH);
-
-      let testPasswordEncoded = '';
-      testPasswordEncoded = await encodeSensitiveTextAsync({
-        text: testPasswordRaw,
-        key: encodeSensitiveTextKey,
-      });
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '⚠️ Wrong password',
-          name: 'decodeSensitiveTextAsync(Wrong password) 😃',
-          fn: () =>
-            decodeSensitiveTextAsync({
-              encodedText:
-                'SENSITIVE_ENCODE::AE7EADC1-CDA0-45FA-A340-E93BEDDEA21E::91bd9aee6525991dbde19d4d51f7265904d5db1592c6b0dfcdcc6ecfddd447883b2797a16926ddbea96c80acaaf99c0bee98282c0f966938095e1369da781ca7',
-              key: encodeSensitiveTextKey,
-            }),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: testPasswordRaw,
-          name: 'decodeSensitiveTextAsync 😃',
-          fn: () =>
-            decodeSensitiveTextAsync({
-              encodedText: testPasswordEncoded,
-              key: encodeSensitiveTextKey,
-            }),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: testPasswordRaw,
-          name: 'decodeSensitiveTextAsync(2) 😃',
-          fn: () =>
-            decodeSensitiveTextAsync({
-              encodedText:
-                'SENSITIVE_ENCODE::AE7EADC1-CDA0-45FA-A340-E93BEDDEA21E::4148c9fc99fa20bb83d3c925b6b94f9cd8e1ba45e21ddfc40b9d3627b17adcc9dbaf4805799fd7da5b581ea70bd31b7876d2bbf6a53d6956c2afb17adbbc5f2f',
-              key: encodeSensitiveTextKey,
-            }),
-        }),
-      );
-
-      // console.log('testPasswordEncoded', {
-      //   encryptAsyncTestResult: bufferUtils.bytesToHex(encryptAsyncTestResult),
-      //   // customIv: bufferUtils.bytesToHex(customIv),
-      //   // customIvLength: customIv.length,
-      //   // customSalt: bufferUtils.bytesToHex(customSalt),
-      //   // customSaltLength: customSalt.length,
-      //   testPasswordRaw,
-      //   testPasswordEncoded,
-      //   testPasswordDecoded,
-      //   testPasswordDecoded2,
-      //   // testPasswordDecoded3: testPasswordDecoded3 || '---',
-      // });
-
-      let testPassword = '';
-      testPassword =
-        await backgroundApiProxy.servicePassword.encodeSensitiveText({
-          text: testPasswordRaw,
-        });
-
-      let rawTestPassword = '';
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: testPasswordRaw,
-          name: 'decodeSensitiveText 😃',
-          fn: async () => {
-            rawTestPassword =
-              await backgroundApiProxy.servicePassword.decodeSensitiveText({
-                encodedText: testPassword,
-              });
-            return rawTestPassword;
-          },
-        }),
-      );
-
-      const testSeed: IBip39RevealableSeed = {
-        entropyWithLangPrefixed: '00112233445566778899aabbccddeeff',
-        seed: '00112233445566778899aabbccddeeff00112233445566778899aabbccddeeff',
-      };
-      const curveName: ICurveName = 'secp256k1';
-      const prefix = 'm';
-      const relPaths = ['0/0', '0/1', "44'/0'/0'/0/0"];
-
-      let encryptedSeed = '';
-      encryptedSeed = await encryptRevealableSeed({
-        rs: testSeed,
-        password: testPassword,
-      });
-
-      const batchGetPublicKeysExpect =
-        '039963f5256af9f48d9d4a340e352f8cec3719b48d9f8514a5495785bb7d7bdac3';
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: batchGetPublicKeysExpect,
-          name: 'batchGetPublicKeys(useWebembedApi) 😃',
-          fn: async () => {
-            const r1 = await batchGetPublicKeys({
-              curveName,
-              hdCredential: encryptedSeed,
-              password: testPassword,
-              prefix,
-              relPaths,
-            });
-            return r1?.[2]?.extendedKey?.key;
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: batchGetPublicKeysExpect,
-          name: 'batchGetPublicKeys(useWebembedApi, byAsyncSubCalls) 😃',
-          fn: async () => {
-            const r1 = await batchGetPublicKeys({
-              curveName,
-              hdCredential: encryptedSeed,
-              password: testPassword,
-              prefix,
-              relPaths,
-              byAsyncSubCalls: true,
-            });
-            return r1?.[2]?.extendedKey?.key;
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: batchGetPublicKeysExpect,
-          name: 'batchGetPublicKeys(useRnJsCrypto)',
-          fn: async () => {
-            const r1 = await batchGetPublicKeys({
-              curveName,
-              hdCredential: encryptedSeed,
-              password: testPassword,
-              prefix,
-              relPaths,
-              useWebembedApi: false,
-            });
-            return r1?.[2]?.extendedKey?.key;
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: batchGetPublicKeysExpect,
-          name: 'batchGetPublicKeys(useRnJsCrypto, byAsyncSubCalls)',
-          fn: async () => {
-            const r1 = await batchGetPublicKeys({
-              curveName,
-              hdCredential: encryptedSeed,
-              password: testPassword,
-              prefix,
-              relPaths,
-              byAsyncSubCalls: true,
-              useWebembedApi: false,
-            });
-            return r1?.[2]?.extendedKey?.key;
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            '8ff67563c060ca12aac18757221cea72482d139ea65d5f5d4f55a05c69ae87ebad76f31087e49bbc59ac0f08d679e4c0869f9babcc7ec4c2557a42abebb072b4',
-          name: 'encryptAsync 😃',
-          fn: async () =>
-            encryptAsync({
-              password:
-                await backgroundApiProxy.servicePassword.encodeSensitiveText({
-                  text: testPasswordRaw,
-                }),
-              data: bufferUtils.utf8ToBytes(testPasswordRaw),
-              customIv,
-              customSalt,
-            }),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            '8ff67563c060ca12aac18757221cea72482d139ea65d5f5d4f55a05c69ae87ebad76f31087e49bbc59ac0f08d679e4c0869f9babcc7ec4c2557a42abebb072b4',
-          name: 'encryptAsync(useWebembedApi)',
-          fn: async () => {
-            return encryptAsync({
-              password:
-                await backgroundApiProxy.servicePassword.encodeSensitiveText({
-                  text: testPasswordRaw,
-                }),
-              data: bufferUtils.utf8ToBytes(testPasswordRaw),
-              useWebembedApi: true,
-              customIv,
-              customSalt,
-            });
-          },
-        }),
-      );
-
-      const r6 = await encryptAsync({
-        password: await backgroundApiProxy.servicePassword.encodeSensitiveText({
-          text: testPasswordRaw,
-        }),
-        data: bufferUtils.utf8ToBytes(testPasswordRaw),
-        useWebembedApi: true,
-      });
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '70617373776f7264313233',
-          name: 'decryptAsync 😃',
-          fn: async () =>
-            decryptAsync({
-              password:
-                await backgroundApiProxy.servicePassword.encodeSensitiveText({
-                  text: testPasswordRaw,
-                }),
-              data: r6,
-            }),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '⚠️ Wrong password',
-          name: 'decryptAsync (wrong password) 😃',
-          fn: async () =>
-            decryptAsync({
-              password:
-                await backgroundApiProxy.servicePassword.encodeSensitiveText({
-                  text: `1111`,
-                }),
-              data: r6,
-            }),
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '70617373776f7264313233',
-          name: 'decryptAsync(useWebembedApi)',
-          fn: async () =>
-            decryptAsync({
-              password:
-                await backgroundApiProxy.servicePassword.encodeSensitiveText({
-                  text: testPasswordRaw,
-                }),
-              data: r6,
-              useWebembedApi: true,
-            }),
-        }),
-      );
-
-      const testMnemonic =
-        'test test test test test test test test test test test junk';
-      const rs = mnemonicToRevealableSeed(testMnemonic, 'optional passphrase');
-      const hdCredential = await encryptRevealableSeed({
-        rs,
-        password: testPassword,
-      });
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'test test test test test test test test test test test junk',
-          name: 'mnemonicFromEntropyAsync(useRnJsCrypto)😃',
-          fn: async () => {
-            return mnemonicFromEntropyAsync({
-              hdCredential,
-              password: testPassword,
-            });
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: 'test test test test test test test test test test test junk',
-          name: 'mnemonicFromEntropyAsync(useWebembedApi)',
-          fn: async () => {
-            return mnemonicFromEntropyAsync({
-              hdCredential,
-              password: testPassword,
-              useWebembedApi: true,
-            });
-          },
-        }),
-      );
-
-      const testPassphrase = 'optional passphrase';
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            'bc0d03ab4f8871dd4a7a68423894bb88fb54973899e4721c9dffd09a5b589171b5712b27da764f7be0653ba361f445b4f9251b490525833b644b7a13eebc7e2c',
-          name: 'mnemonicToSeedAsync(useWebembedApi)😃',
-          fn: async () => {
-            return mnemonicToSeedAsync({
-              mnemonic: testMnemonic,
-              passphrase: testPassphrase,
-            });
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect:
-            'bc0d03ab4f8871dd4a7a68423894bb88fb54973899e4721c9dffd09a5b589171b5712b27da764f7be0653ba361f445b4f9251b490525833b644b7a13eebc7e2c',
-          name: 'mnemonicToSeedAsync(useRnJsCrypto)',
-          fn: async () => {
-            return mnemonicToSeedAsync({
-              mnemonic: testMnemonic,
-              passphrase: testPassphrase,
-              useWebembedApi: false,
-            });
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '045a91ef',
-          name: 'generateRootFingerprintHexAsync(useRnJsCrypto)😃',
-          fn: async () => {
-            return generateRootFingerprintHexAsync({
-              curveName: 'secp256k1',
-              hdCredential,
-              password: testPassword,
-            });
-          },
-        }),
-      );
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '045a91ef',
-          name: 'generateRootFingerprintHexAsync(useWebembedApi)',
-          fn: async () => {
-            return generateRootFingerprintHexAsync({
-              curveName: 'secp256k1',
-              hdCredential,
-              password: testPassword,
-              useWebembedApi: true,
-            });
-          },
-        }),
-      );
-
-      setResult(
-        stringUtils.stableStringify(
-          tasks,
-          // null,
-          stringUtils.STRINGIFY_REPLACER.bufferToHex,
-          2,
-        ),
-      );
-    } catch (error) {
-      Toast.error({
-        title: `SecretFunctions failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  const testSecretFunctions2 = async () => {
-    try {
-      const {
-        encryptRevealableSeed,
-        generateRootFingerprintHexAsync,
-        mnemonicToRevealableSeed,
-      } = await loadCoreSecret();
-      const tasks: IRunAppCryptoTestTaskResult[] = [];
-      const testPasswordRaw = 'password123';
-
-      let testPassword = '';
-      testPassword =
-        await backgroundApiProxy.servicePassword.encodeSensitiveText({
-          text: testPasswordRaw,
-        });
-
-      const testMnemonic =
-        'test test test test test test test test test test test junk';
-      const rs = mnemonicToRevealableSeed(testMnemonic, 'optional passphrase');
-      const hdCredential = await encryptRevealableSeed({
-        rs,
-        password: testPassword,
-      });
-
-      tasks.push(
-        await runAppCryptoTestTask({
-          expect: '045a91ef',
-          name: 'generateRootFingerprintHexAsync(useRnJsCrypto)😃',
-          fn: async () => {
-            return generateRootFingerprintHexAsync({
-              curveName: 'secp256k1',
-              hdCredential,
-              password: testPassword,
-            });
-          },
-        }),
-      );
-
-      setResult(
-        stringUtils.stableStringify(
-          tasks,
-          // null,
-          stringUtils.STRINGIFY_REPLACER.bufferToHex,
-          2,
-        ),
-      );
-    } catch (error) {
-      Toast.error({
-        title: `SecretFunctions2 failed: ${(error as Error).message}`,
-      });
-    }
-  };
-
-  return (
-    <PartContainer title="SecretFunctions Test">
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testSecretFunctions)}
-      >
-        Test SecretFunctions
-      </Button>
-      <Button
-        variant="primary"
-        onPress={() => runCryptoGalleryTestExclusive(testSecretFunctions2)}
-      >
-        Test SecretFunctions2
-      </Button>
-      {result ? <SizableText size="$bodyMd">{result}</SizableText> : null}
-    </PartContainer>
-  );
-}
-
-function JotaiDemoPriceInfo() {
-  const [demoPriceInfo, setDemoPriceInfo] = useDemoPriceInfoAtom();
-
-  return (
-    <DebugRenderTracker>
-      <Stack>
-        <SizableText size="$bodyMd">
-          {JSON.stringify(demoPriceInfo, null, 2)}
-        </SizableText>
-        <Button
-          variant="primary"
-          onPress={() =>
-            setDemoPriceInfo((prev) => ({ ...prev, price: 10, info: 'info' }))
-          }
-        >
-          setDemoPriceInfo(new object)
-        </Button>
-        <Button
-          variant="primary"
-          onPress={() => setDemoPriceInfo((prev) => prev)}
-        >
-          setDemoPriceInfo(prev object)
-        </Button>
-      </Stack>
-    </DebugRenderTracker>
-  );
-}
 
 const CryptoGallery = () => (
   <Layout
@@ -4398,6 +31,9 @@ const CryptoGallery = () => (
             <CustomAccordion>
               <CustomAccordionItem title="AES-GCM PBKDF2 (v2 test)">
                 <AESGcmV2Test />
+              </CustomAccordionItem>
+              <CustomAccordionItem title="ADA Address Perf">
+                <AdaAddressPerfTest />
               </CustomAccordionItem>
               <CustomAccordionItem title="Secret Functions Crypto Perf (v2 test)">
                 <NativeWebEmbedCryptoPerfTest />
